Split session tests by subsystem
