+++ crates/truthdb-core/src/session/mod.rs
@@ -2125,3918 +2125,4 @@ fn lock_timeout_error() -> SqlError {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::engine::Engine;
-    use crate::engine::StatementResult;
-    use crate::relstore::types::Datum;
-    use crate::storage::{Storage, StorageOptions};
-    use std::collections::{HashMap, HashSet};
-    use std::path::PathBuf;
-
-    fn graph(edges: &[(u64, &[u64])]) -> HashMap<u64, HashSet<u64>> {
-        edges
-            .iter()
-            .map(|(from, to)| (*from, to.iter().copied().collect()))
-            .collect()
-    }
-
-    #[test]
-    fn find_cycle_detects_and_ignores_cycles() {
-        // No edges / acyclic chain / DAG -> None.
-        assert!(find_cycle(&graph(&[])).is_none());
-        assert!(find_cycle(&graph(&[(1, &[2]), (2, &[3]), (3, &[])])).is_none());
-        assert!(find_cycle(&graph(&[(1, &[2, 3]), (2, &[3]), (3, &[])])).is_none());
-
-        // A 2-cycle where the second node is reached as a neighbor before it is
-        // colored — the case that regressed when unvisited nodes defaulted to
-        // "done" instead of "unvisited".
-        let c2 = find_cycle(&graph(&[(1, &[2]), (2, &[1])])).expect("2-cycle");
-        assert_eq!(c2.iter().copied().collect::<HashSet<_>>(), [1, 2].into());
-
-        // A 3-cycle with a dead-end branch (4 holds a lock but is not waiting).
-        let c3 = find_cycle(&graph(&[(1, &[2, 4]), (2, &[3]), (3, &[1])])).expect("3-cycle");
-        assert_eq!(c3.iter().copied().collect::<HashSet<_>>(), [1, 2, 3].into());
-
-        // A self-loop (a transaction waiting on itself should never happen, but
-        // the detector must not miss it).
-        assert!(find_cycle(&graph(&[(1, &[1])])).is_some());
-    }
-
-    fn test_storage_options() -> StorageOptions {
-        StorageOptions {
-            size_gib: 1,
-            wal_ratio: 0.05,
-            metadata_ratio: 0.08,
-            snapshot_ratio: 0.02,
-            allocator_ratio: 0.02,
-            reserved_ratio: 0.17,
-            default_collation: None,
-        }
-    }
-
-    fn unique_temp_path(label: &str) -> PathBuf {
-        let mut path = std::env::temp_dir();
-        let nanos = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .expect("system time before unix epoch")
-            .as_nanos();
-        path.push(format!("truthdb-lock-{label}-{nanos}.db"));
-        path
-    }
-
-    /// A running engine plus the temp file backing it (removed on drop).
-    struct Harness {
-        handle: EngineHandle,
-        path: PathBuf,
-    }
-
-    impl Drop for Harness {
-        fn drop(&mut self) {
-            self.handle.shutdown();
-            let _ = std::fs::remove_file(&self.path);
-        }
-    }
-
-    fn start(timeout: Duration) -> Harness {
-        let path = unique_temp_path("engine");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let engine = Engine::new(storage).expect("engine");
-        let (handle, _join) = spawn_engine_with_timeout(engine, timeout);
-        Harness { handle, path }
-    }
-
-    /// A single-worker harness. With one worker the loop's wait is the only
-    /// thing deciding when the sweep runs; with several, a sibling that
-    /// snapshotted an earlier deadline can wake and sweep on another's behalf,
-    /// masking a wait that is too long.
-    fn start_single_worker(idle: Option<Duration>) -> Harness {
-        let path = unique_temp_path("engine-1worker");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let engine = Engine::new(storage).expect("engine");
-        let (handle, _join) = spawn_engine_pool(engine, LOCK_WAIT_TIMEOUT, idle, 1);
-        Harness { handle, path }
-    }
-
-    /// A harness whose idle-transaction reaper fires after `idle` (or never,
-    /// when `None`), so the reaper is testable without a real 10 min wait.
-    fn start_with_idle(idle: Option<Duration>) -> Harness {
-        let path = unique_temp_path("engine-idle");
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let engine = Engine::new(storage).expect("engine");
-        let (handle, _join) = spawn_engine_with_idle_timeout(engine, idle);
-        Harness { handle, path }
-    }
-
-    /// The `id` column (column 0) of the first rowset, as i64s.
-    fn ids(reply: &BatchReply) -> Vec<i64> {
-        for result in &reply.outcome.results {
-            if let StatementResult::Rows(rowset) = result {
-                return rowset
-                    .rows
-                    .iter()
-                    .map(|row| match row[0] {
-                        Datum::TinyInt(v) => v as i64,
-                        Datum::SmallInt(v) => v as i64,
-                        Datum::Int(v) => v as i64,
-                        Datum::BigInt(v) => v,
-                        ref other => panic!("expected integer id, got {other:?}"),
-                    })
-                    .collect();
-            }
-        }
-        panic!("no rowset in outcome");
-    }
-
-    fn error_number(reply: &BatchReply) -> Option<i32> {
-        reply.outcome.error.as_ref().map(|e| e.number)
-    }
-
-    // ---- idle-transaction reaper ----------------------------------------
-
-    #[tokio::test]
-    async fn idle_transaction_is_reaped_and_its_locks_released() {
-        // A client that opens a transaction and goes silent without
-        // disconnecting must not hold its locks forever.
-        let h = start_with_idle(Some(Duration::from_millis(150)));
-        let a = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        // Session A takes a write lock and then abandons the transaction.
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (1);".into())
-            .await
-            .unwrap();
-
-        // B blocks on A's lock while A still holds it, so wait for the reaper.
-        // Once it fires, A's write is rolled back and B sees an empty table.
-        let reply = h
-            .handle
-            .run_batch(b, "SELECT id FROM t".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            None,
-            "B must proceed once the abandoned transaction is reaped: {:?}",
-            reply.outcome.error
-        );
-        assert_eq!(ids(&reply), Vec::<i64>::new(), "the reaped write is undone");
-
-        // A is told its transaction was reaped rather than left to discover it
-        // at a COMMIT that fails for a confusing reason.
-        let reply = h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        assert_eq!(error_number(&reply), Some(1205));
-        // The signal fires once; the now-transactionless COMMIT then reports the
-        // ordinary 3902.
-        let reply = h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        assert_eq!(error_number(&reply), Some(3902));
-    }
-
-    /// Drains a batch's event stream through its terminal event.
-    async fn drain_events(mut rx: mpsc::UnboundedReceiver<BatchEvent>) -> Vec<BatchEvent> {
-        let mut events = Vec::new();
-        while let Some(event) = rx.recv().await {
-            let terminal = matches!(event, BatchEvent::Complete { .. } | BatchEvent::Failed(_));
-            events.push(event);
-            if terminal {
-                break;
-            }
-        }
-        events
-    }
-
-    fn no_cancel() -> Arc<AtomicBool> {
-        Arc::new(AtomicBool::new(false))
-    }
-
-    /// The `in_transaction` flag of every StatementDone, in stream order.
-    fn done_flags(events: &[BatchEvent]) -> Vec<bool> {
-        events
-            .iter()
-            .filter_map(|event| match event {
-                BatchEvent::StatementDone { in_transaction, .. } => Some(*in_transaction),
-                _ => None,
-            })
-            .collect()
-    }
-
-    #[tokio::test]
-    async fn statement_dones_carry_their_own_transaction_state() {
-        // Each DONE reports the transaction state after *its own* statement —
-        // BEGIN and the SELECT inside the transaction say so, the COMMIT says
-        // it ended — instead of the batch's final state stamped on all three.
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let events = drain_events(h.handle.stream_batch(
-            s,
-            "BEGIN TRANSACTION; SELECT 1 AS one; COMMIT".into(),
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(done_flags(&events), [true, true, false]);
-        assert!(
-            matches!(
-                events.last(),
-                Some(BatchEvent::Complete {
-                    in_transaction: false
-                })
-            ),
-            "Complete carries the batch-final state: {events:?}"
-        );
-    }
-
-    /// Fills `t` with `1..=n` single-column PK rows.
-    async fn fill(h: &Harness, s: SessionId, n: usize) {
-        h.handle
-            .run_batch(s, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        // Batched into 500-tuple inserts: the per-expression node budget
-        // keeps giant single statements bounded, and chunks stay far from
-        // every limit.
-        let ids: Vec<usize> = (1..=n).collect();
-        for chunk in ids.chunks(500) {
-            let values: Vec<String> = chunk.iter().map(|i| format!("({i})")).collect();
-            let reply = h
-                .handle
-                .run_batch(s, format!("INSERT INTO t VALUES {}", values.join(", ")))
-                .await
-                .unwrap();
-            assert!(reply.outcome.error.is_none(), "{:?}", reply.outcome.error);
-        }
-    }
-
-    #[tokio::test]
-    async fn a_mid_scan_error_keeps_the_rows_already_streamed() {
-        // A streamed SELECT that fails part-way has already emitted the rows
-        // that preceded the failure — rows leave while the statement is still
-        // running. The buffered path emitted nothing for a failed statement,
-        // so any Columns/Rows here prove the stream is real.
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        fill(&h, s, 600).await;
-        // The WHERE divides by zero at id = 600, after 599 kept rows: two full
-        // 256-row chunks are already out, the partial third is dropped.
-        let events = drain_events(h.handle.stream_batch(
-            s,
-            "SELECT id FROM t WHERE 10 / (id - 600) > -100".into(),
-            no_cancel(),
-        ))
-        .await;
-        let streamed: usize = events
-            .iter()
-            .map(|event| match event {
-                BatchEvent::Rows(rows) => rows.len(),
-                _ => 0,
-            })
-            .sum();
-        assert_eq!(streamed, 512, "two full chunks precede the failure");
-        assert!(
-            events
-                .iter()
-                .any(|e| matches!(e, BatchEvent::Error(err) if err.number == 8134)),
-            "the divide-by-zero still reaches the client: {events:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn a_caught_mid_scan_error_closes_the_open_rowset() {
-        // A TRY/CATCH swallows the error, but the failed SELECT's result set
-        // had already started streaming — it must be closed (StatementAborted)
-        // before the CATCH's own result set opens, and no Error event follows.
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        fill(&h, s, 300).await;
-        let events = drain_events(
-            h.handle.stream_batch(
-                s,
-                "BEGIN TRY SELECT id FROM t WHERE 10 / (id - 300) > -100 END TRY \
-             BEGIN CATCH SELECT 99 AS caught END CATCH"
-                    .into(),
-                no_cancel(),
-            ),
-        )
-        .await;
-        let aborted = events
-            .iter()
-            .position(|e| matches!(e, BatchEvent::StatementAborted { .. }))
-            .expect("the failed SELECT's rowset is closed");
-        let caught = events
-            .iter()
-            .position(
-                |e| matches!(e, BatchEvent::Columns(cols) if cols.first().is_some_and(|c| c.name == "caught")),
-            )
-            .expect("the CATCH's rowset follows");
-        assert!(aborted < caught, "close before reopening: {events:?}");
-        assert!(
-            !events.iter().any(|e| matches!(e, BatchEvent::Error(_))),
-            "a caught error never surfaces: {events:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn a_continued_mid_scan_error_closes_the_rowset_and_reports_last() {
-        // Under XACT_ABORT OFF a non-fatal in-transaction error rolls back only
-        // its statement and the batch continues — so the half-streamed rowset
-        // closes, the following statements run, and the error is reported at
-        // the end of the batch, exactly where the buffered path put it.
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        fill(&h, s, 300).await;
-        let events = drain_events(
-            h.handle.stream_batch(
-                s,
-                "BEGIN TRANSACTION; SELECT id FROM t WHERE 10 / (id - 300) > -100; \
-             SELECT 7 AS after; COMMIT"
-                    .into(),
-                no_cancel(),
-            ),
-        )
-        .await;
-        assert!(
-            events.iter().any(|e| matches!(
-                e,
-                BatchEvent::StatementAborted {
-                    in_transaction: true
-                }
-            )),
-            "the failed SELECT's rowset closes, still in-transaction: {events:?}"
-        );
-        // BEGIN, the surviving SELECT, then COMMIT — each DONE with its own state.
-        assert_eq!(done_flags(&events), [true, true, false]);
-        let error = events
-            .iter()
-            .position(|e| matches!(e, BatchEvent::Error(err) if err.number == 8134))
-            .expect("the continued error is still reported");
-        assert_eq!(error, events.len() - 2, "after every result: {events:?}");
-    }
-
-    #[tokio::test]
-    async fn the_sweep_runs_on_time_even_with_a_batch_parked_further_out() {
-        // A worker must not sleep past the sweep just because the nearest parked
-        // deadline is further out: the abandoned transaction the parked batch is
-        // waiting on would not be reaped until that deadline, and the waiter
-        // would die of its own timeout (1205) first — the very lock it was
-        // waiting for having been reclaimable the whole time.
-        //
-        // Single-worker, because with several workers a sibling holding an
-        // earlier deadline snapshot can wake and sweep anyway, hiding this.
-        let h = start_single_worker(Some(Duration::from_millis(150)));
-        let a = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        // A abandons a transaction holding a write lock.
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (1);".into())
-            .await
-            .unwrap();
-        // B parks on A's lock with a 5 s deadline — far beyond the 150 ms sweep.
-        let reply = h
-            .handle
-            .run_batch(b, "SELECT id FROM t".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            None,
-            "B must be unblocked by the sweep, not killed at its own deadline: {:?}",
-            reply.outcome.error
-        );
-        assert_eq!(ids(&reply), Vec::<i64>::new(), "A's write was rolled back");
-    }
-
-    #[tokio::test]
-    async fn active_transaction_is_not_reaped() {
-        // The reaper must only touch *idle* sessions: a session that keeps
-        // working keeps its transaction, however long it stays open.
-        let h = start_with_idle(Some(Duration::from_millis(150)));
-        let a = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (1);".into())
-            .await
-            .unwrap();
-        // Keep touching the session across more than the idle timeout.
-        for i in 2..=6 {
-            tokio::time::sleep(Duration::from_millis(50)).await;
-            let reply = h
-                .handle
-                .run_batch(a, format!("INSERT INTO t VALUES ({i})"))
-                .await
-                .unwrap();
-            assert_eq!(
-                error_number(&reply),
-                None,
-                "an active session must keep its transaction"
-            );
-        }
-        // The transaction survived and commits everything it wrote.
-        let reply = h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        let reply = h
-            .handle
-            .run_batch(a, "SELECT id FROM t ORDER BY id".into())
-            .await
-            .unwrap();
-        assert_eq!(ids(&reply), vec![1, 2, 3, 4, 5, 6]);
-    }
-
-    #[tokio::test]
-    async fn idle_reaper_can_be_disabled() {
-        // With the reaper off, an idle transaction is left alone.
-        let h = start_with_idle(None);
-        let a = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (1);".into())
-            .await
-            .unwrap();
-        tokio::time::sleep(Duration::from_millis(300)).await;
-        // The transaction is still open and still commits.
-        let reply = h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        let reply = h
-            .handle
-            .run_batch(a, "SELECT id FROM t".into())
-            .await
-            .unwrap();
-        assert_eq!(ids(&reply), vec![1]);
-    }
-
-    /// A bare Scheduler + Engine, for pinning the sweep's guards directly
-    /// instead of racing real timers.
-    fn bare(label: &str, idle: Option<Duration>) -> (Engine, Scheduler, PathBuf) {
-        let path = unique_temp_path(label);
-        let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-        let engine = Engine::new(storage).expect("engine");
-        (engine, Scheduler::new(LOCK_WAIT_TIMEOUT, idle), path)
-    }
-
-    #[test]
-    fn a_running_batch_is_never_reaped_however_idle_the_session_looks() {
-        // The reaper's whole safety argument: while a batch runs its context has
-        // been moved out by take_ctx, so the session reports no open
-        // transaction and the sweep cannot select it. Pinned directly, with a
-        // zero idle timeout so that guard is the *only* thing protecting it.
-        let (engine, mut sched, path) = bare("reap-running", Some(Duration::ZERO));
-        let id = sched.sessions.open(
-            "truthdb".into(),
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "sa".into(),
-            0,
-            "dbo".into(),
-            0,
-        );
-        {
-            let ctx = &mut sched.sessions.get_mut(id).expect("session").txn_ctx;
-            engine
-                .sql_batch("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)", ctx)
-                .expect("create");
-            engine
-                .sql_batch("BEGIN TRAN; INSERT INTO t VALUES (1);", ctx)
-                .expect("begin");
-        }
-        assert!(
-            sched
-                .sessions
-                .get(id)
-                .expect("session")
-                .txn_ctx
-                .has_open_transaction(),
-            "precondition: the session holds an open transaction"
-        );
-
-        // Simulate the batch being dispatched: the context moves to the worker.
-        let ctx = sched.take_ctx(id);
-        assert!(
-            !sched
-                .sessions
-                .get(id)
-                .expect("session")
-                .txn_ctx
-                .has_open_transaction(),
-            "take_ctx must leave the session reporting no open transaction"
-        );
-        assert!(
-            !sched.reap_idle_txns(&engine),
-            "a session whose batch is running must never be reaped"
-        );
-
-        // Restoring it makes the very same session reapable — proving the test
-        // above is not passing merely because nothing is ever reapable.
-        sched.finish(&engine, id, ctx);
-        assert!(
-            sched.reap_idle_txns(&engine),
-            "once idle again, the transaction is reaped"
-        );
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn a_session_with_a_parked_batch_is_not_reaped() {
-        // A parked batch is waiting on locks, not abandoned; its own deadline
-        // reaps it. Reaping its transaction underneath it would run the batch
-        // against a rolled-back transaction.
-        let (engine, mut sched, path) = bare("reap-parked", Some(Duration::ZERO));
-        let id = sched.sessions.open(
-            "truthdb".into(),
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "sa".into(),
-            0,
-            "dbo".into(),
-            0,
-        );
-        {
-            let ctx = &mut sched.sessions.get_mut(id).expect("session").txn_ctx;
-            engine
-                .sql_batch("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)", ctx)
-                .expect("create");
-            engine
-                .sql_batch("BEGIN TRAN; INSERT INTO t VALUES (1);", ctx)
-                .expect("begin");
-        }
-        let (tx, _rx) = mpsc::unbounded_channel();
-        let reply = BatchSink::new(tx);
-        sched.parked.push_back(Parked {
-            session: id,
-            sql: "SELECT id FROM t".into(),
-            params: Vec::new(),
-            proc_tail: None,
-            cancel: Arc::new(AtomicBool::new(false)),
-            reply,
-            needs: Vec::new(),
-            deadline: Instant::now() + Duration::from_secs(5),
-            epoch: 0,
-        });
-        assert!(
-            !sched.reap_idle_txns(&engine),
-            "a session with a parked batch must not be reaped"
-        );
-
-        // Drop the parked entry and the same session is reaped — the guard, not
-        // some other condition, is what protected it.
-        sched.parked.clear();
-        assert!(
-            sched.reap_idle_txns(&engine),
-            "with nothing parked, the idle transaction is reaped"
-        );
-        let _ = std::fs::remove_file(path);
-    }
-
-    /// Adversarial review probe: ALTER PROCEDURE bumps the lock-analysis
-    /// epoch, so a batch parked with a lock set analyzed over the OLD body is
-    /// re-analyzed at grant time against the NEW body — otherwise the new
-    /// body's writes would run under the old body's read locks (the stale
-    /// twice-derived-decision class from the Stage 13 review).
-    #[test]
-    fn review_poc_alter_procedure_reanalyzes_a_parked_exec() {
-        let (engine, mut sched, path) = bare("epoch-proc-alter", None);
-        let blocker = sched.sessions.open(
-            "truthdb".into(),
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "sa".into(),
-            0,
-            "dbo".into(),
-            0,
-        );
-        let waiter = sched.sessions.open(
-            "truthdb".into(),
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "sa".into(),
-            0,
-            "dbo".into(),
-            0,
-        );
-        {
-            let mut ctx = crate::engine::TxnContext::default();
-            engine
-                .sql_batch("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)", &mut ctx)
-                .expect("create table");
-            engine
-                .sql_batch("CREATE PROCEDURE p AS SELECT id FROM t", &mut ctx)
-                .expect("create proc");
-        }
-        // The blocker holds Database X so nothing becomes grantable.
-        assert!(sched.try_acquire(
-            blocker.raw(),
-            &[(Resource::Database, LockMode::Exclusive)],
-            true
-        ));
-        // Park an EXEC analyzed against the OLD (read-only) body.
-        let epoch = engine.lock_analysis_epoch();
-        let needs = engine.analyze_locks(
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "EXEC p",
-            crate::engine::Isolation::ReadCommitted,
-        );
-        assert!(
-            !needs
-                .iter()
-                .any(|(_, m)| matches!(m, LockMode::IntentExclusive | LockMode::Exclusive)),
-            "precondition: the old body takes no write locks: {needs:?}"
-        );
-        let (tx, _rx) = mpsc::unbounded_channel();
-        sched.parked.push_back(Parked {
-            session: waiter,
-            sql: "EXEC p".into(),
-            params: Vec::new(),
-            proc_tail: None,
-            cancel: Arc::new(AtomicBool::new(false)),
-            reply: BatchSink::new(tx),
-            needs,
-            deadline: Instant::now() + Duration::from_secs(5),
-            epoch,
-        });
-        // The body is replaced while the batch is parked.
-        {
-            let mut ctx = crate::engine::TxnContext::default();
-            engine
-                .sql_batch("ALTER PROCEDURE p AS INSERT INTO t VALUES (1)", &mut ctx)
-                .expect("alter proc");
-        }
-        // The grant path must re-analyze (the blocker still prevents the
-        // grant itself, so the refreshed needs stay inspectable).
-        assert!(
-            sched.next_grantable(&engine).is_none(),
-            "Database X is still held; nothing may be granted"
-        );
-        let refreshed = &sched.parked.front().expect("still parked").needs;
-        assert!(
-            refreshed
-                .iter()
-                .any(|(_, m)| matches!(m, LockMode::IntentExclusive | LockMode::Exclusive)),
-            "the parked EXEC's lock set was re-analyzed against the NEW \
-             body (INSERT): {refreshed:?}"
-        );
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[tokio::test]
-    async fn the_pool_actually_spawns_a_maintenance_thread() {
-        // `housekeeping_runs_with_no_worker_free_to_do_it` builds `Shared` by
-        // hand and starts the thread itself, so it would pass just as happily
-        // if the supervisor never spawned one. This covers the wiring: run the
-        // real `spawn_engine` path and wait for a maintenance thread to report
-        // in. (A sibling test's pool satisfying this is fine — it is the same
-        // supervisor code either way; what fails is nobody spawning one at all.)
-        let h = start_with_idle(Some(Duration::from_millis(150)));
-        let deadline = Instant::now() + Duration::from_secs(5);
-        while MAINTENANCE_STARTS.load(Ordering::Relaxed) == 0 && Instant::now() < deadline {
-            tokio::time::sleep(Duration::from_millis(10)).await;
-        }
-        assert!(
-            MAINTENANCE_STARTS.load(Ordering::Relaxed) > 0,
-            "the pool spawns a maintenance thread"
-        );
-        drop(h);
-    }
-
-    #[test]
-    fn both_ways_of_shutting_the_pool_down_stop_every_thread() {
-        // The server calls `shutdown` and then joins; the tests just drop the
-        // handle. Only the first closes the inbox by itself — the second relies
-        // on the handle token, since the `Arc<Inbox>` the workers hold can never
-        // reach zero. Joining the supervisor is what proves it: it joins the
-        // workers and the maintenance thread first, so it only returns if every
-        // one of them noticed.
-        for explicit in [true, false] {
-            let path = unique_temp_path("shutdown");
-            let storage = Storage::create(path.clone(), test_storage_options()).expect("create");
-            let engine = Engine::new(storage).expect("engine");
-            let (handle, join) = spawn_engine(engine);
-            if explicit {
-                handle.shutdown();
-                drop(handle);
-            } else {
-                drop(handle);
-            }
-            join.join().expect("the pool shut down");
-            let _ = std::fs::remove_file(path);
-        }
-    }
-
-    #[tokio::test]
-    async fn a_batch_the_idle_reaper_unblocks_is_handed_to_a_worker() {
-        // The reaper runs off-worker now, so releasing a lock and running what
-        // that unblocks happen on different threads. Workers block on the inbox
-        // indefinitely — there is no timeout to fall back on — so a reap that
-        // did not nudge would leave this batch parked forever rather than late.
-        let h = start_with_idle(Some(Duration::from_millis(150)));
-        let a = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        // A abandons a write lock; B parks behind it and can only be rescued by
-        // the maintenance thread reaping A and then nudging a worker.
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (1);".into())
-            .await
-            .unwrap();
-        let reply = tokio::time::timeout(
-            Duration::from_secs(10),
-            h.handle.run_batch(b, "SELECT id FROM t".into()),
-        )
-        .await
-        .expect("the rescued batch was handed to a worker, not left parked")
-        .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        assert_eq!(ids(&reply), Vec::<i64>::new(), "the reaped write is undone");
-    }
-
-    #[test]
-    fn the_maintenance_thread_sleeps_between_sweeps_whatever_is_parked() {
-        // An expired waiter that `reap_expired` must NOT reap (its locks are
-        // free, so it is queued for a worker rather than blocked) keeps a
-        // deadline in the past for as long as it sits there. A sweeper that
-        // derived its sleep from that deadline would compute zero and spin,
-        // taking the scheduler mutex thousands of times a second — and would do
-        // it precisely while every worker was busy, which is the case this
-        // thread exists for.
-        let (engine, mut sched, path) = bare("no-spin", Some(Duration::from_millis(50)));
-        let id = sched.sessions.open(
-            "truthdb".into(),
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "sa".into(),
-            0,
-            "dbo".into(),
-            0,
-        );
-        let (tx, _rx) = mpsc::unbounded_channel();
-        let reply = BatchSink::new(tx);
-        sched.parked.push_back(Parked {
-            session: id,
-            sql: "SELECT 1".into(),
-            params: Vec::new(),
-            proc_tail: None,
-            cancel: Arc::new(AtomicBool::new(false)),
-            reply,
-            needs: vec![(Resource::Table(1), LockMode::Shared)],
-            deadline: Instant::now() - Duration::from_secs(60),
-            epoch: 0,
-        });
-        let shared = Arc::new(Shared {
-            engine: Arc::new(engine),
-            scheduler: Mutex::new(sched),
-            inbox: Arc::new(Inbox::new()),
-            stop: AtomicBool::new(false),
-            idle: Mutex::new(()),
-            wake: Condvar::new(),
-            sweeps: std::sync::atomic::AtomicUsize::new(0),
-        });
-        let keeper = Arc::clone(&shared);
-        let maintenance = std::thread::spawn(move || maintenance_loop(&keeper));
-        std::thread::sleep(Duration::from_millis(200));
-        let sweeps = shared.sweeps.load(Ordering::Relaxed);
-        {
-            let _idle = shared.idle.lock().expect("idle mutex poisoned");
-            shared.stop.store(true, Ordering::Release);
-        }
-        shared.wake.notify_all();
-        maintenance.join().expect("maintenance thread");
-
-        // 200ms at a 50ms interval is ~4 sweeps. A spin measures in thousands,
-        // so the bound is loose enough not to be a timing test.
-        assert!(
-            sweeps <= 20,
-            "the thread slept between sweeps; it ran {sweeps} in 200ms"
-        );
-        assert!(
-            shared
-                .scheduler
-                .lock()
-                .expect("scheduler poisoned")
-                .parked
-                .len()
-                == 1,
-            "and it left the grantable waiter for a worker, rather than reaping it"
-        );
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn housekeeping_runs_with_no_worker_free_to_do_it() {
-        // The reapers are the engine's safety valves and must not be hostage to
-        // the pool: a worker only sweeps between batches, so a busy pool used to
-        // mean no sweep. Pinned in its strongest form — a pool with *no workers
-        // at all*, which no amount of load can distinguish itself from — where
-        // only the maintenance thread can do it.
-        let (engine, mut sched, path) = bare("maintenance", Some(Duration::from_millis(50)));
-        let id = sched.sessions.open(
-            "truthdb".into(),
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "sa".into(),
-            0,
-            "dbo".into(),
-            0,
-        );
-        {
-            let ctx = &mut sched.sessions.get_mut(id).expect("session").txn_ctx;
-            engine
-                .sql_batch("CREATE TABLE t (id INT NOT NULL PRIMARY KEY)", ctx)
-                .expect("create");
-            engine
-                .sql_batch("BEGIN TRAN; INSERT INTO t VALUES (1);", ctx)
-                .expect("begin");
-            assert!(
-                ctx.has_open_transaction(),
-                "the session starts with an open txn"
-            );
-        }
-        let shared = Arc::new(Shared {
-            engine: Arc::new(engine),
-            scheduler: Mutex::new(sched),
-            inbox: Arc::new(Inbox::new()),
-            stop: AtomicBool::new(false),
-            idle: Mutex::new(()),
-            wake: Condvar::new(),
-            sweeps: std::sync::atomic::AtomicUsize::new(0),
-        });
-        let keeper = Arc::clone(&shared);
-        let maintenance = std::thread::spawn(move || maintenance_loop(&keeper));
-
-        let deadline = Instant::now() + Duration::from_secs(5);
-        let reaped = loop {
-            {
-                let sched = shared.scheduler.lock().expect("scheduler poisoned");
-                if !sched
-                    .sessions
-                    .get(id)
-                    .expect("session")
-                    .txn_ctx
-                    .has_open_transaction()
-                {
-                    break true;
-                }
-            }
-            if Instant::now() > deadline {
-                break false;
-            }
-            std::thread::sleep(Duration::from_millis(10));
-        };
-        {
-            let _idle = shared.idle.lock().expect("idle mutex poisoned");
-            shared.stop.store(true, Ordering::Release);
-        }
-        shared.wake.notify_all();
-        maintenance.join().expect("maintenance thread");
-        assert!(
-            reaped,
-            "the idle transaction was reaped with no worker to do it"
-        );
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[test]
-    fn a_parked_batch_whose_locks_are_free_is_never_reaped_as_a_victim() {
-        // The deadline is a backstop for a batch stuck behind someone else's
-        // lock. A waiter whose locks are free is not stuck — it is queued for a
-        // worker — so reaping it would report a conflict (1205) that never
-        // existed.
-        let (engine, mut sched, path) = bare("reap-grantable", None);
-        let id = sched.sessions.open(
-            "truthdb".into(),
-            crate::relstore::catalog::DEFAULT_DATABASE_ID,
-            "sa".into(),
-            0,
-            "dbo".into(),
-            0,
-        );
-        let (tx, _rx) = mpsc::unbounded_channel();
-        let reply = BatchSink::new(tx);
-        // Parked, deadline long gone, and nothing holds the lock it wants.
-        sched.parked.push_back(Parked {
-            session: id,
-            sql: "SELECT 1".into(),
-            params: Vec::new(),
-            proc_tail: None,
-            cancel: Arc::new(AtomicBool::new(false)),
-            reply,
-            needs: vec![(Resource::Table(1), LockMode::Shared)],
-            deadline: Instant::now() - Duration::from_secs(60),
-            epoch: 0,
-        });
-        sched.reap_expired(&engine);
-        assert_eq!(
-            sched.parked.len(),
-            1,
-            "an expired waiter whose locks are free is not a deadlock victim"
-        );
-
-        // The same waiter, once something actually blocks it, is reaped — so it
-        // is grantability doing the work above, not a dead reaper.
-        sched
-            .locks
-            .grant(999, Resource::Table(1), LockMode::Exclusive);
-        sched.reap_expired(&engine);
-        assert!(
-            sched.parked.is_empty(),
-            "an expired waiter that is genuinely blocked is still the victim"
-        );
-        let _ = std::fs::remove_file(path);
-    }
-
-    #[tokio::test]
-    async fn reaped_session_is_told_instead_of_silently_autocommitting() {
-        // A client that comes back believing it is still in a transaction must
-        // not have its statements silently autocommit.
-        let h = start_with_idle(Some(Duration::from_millis(150)));
-        let a = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (1);".into())
-            .await
-            .unwrap();
-        tokio::time::sleep(Duration::from_millis(400)).await;
-
-        // The next batch is told the transaction is gone, and does not run.
-        let reply = h
-            .handle
-            .run_batch(a, "INSERT INTO t VALUES (2)".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(1205),
-            "the reap must be reported, not swallowed"
-        );
-
-        // The signal fires once; the session is usable again afterwards.
-        let reply = h
-            .handle
-            .run_batch(a, "SELECT id FROM t".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None);
-        assert_eq!(
-            ids(&reply),
-            Vec::<i64>::new(),
-            "the reaped write is undone, and the rejected INSERT never applied"
-        );
-    }
-
-    #[tokio::test]
-    async fn a_reaped_transaction_leaves_no_savepoints_behind() {
-        // A savepoint holds the undo-log offset of the transaction that recorded
-        // it. One surviving a reap would let ROLLBACK TRANSACTION find a stale
-        // entry in the session's NEXT transaction and hand a dead offset to the
-        // undo log — silently discarding committed work, or panicking.
-        let h = start_with_idle(Some(Duration::from_millis(150)));
-        let a = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(
-                a,
-                "BEGIN TRAN; INSERT INTO t VALUES (1); SAVE TRANSACTION sp1;".into(),
-            )
-            .await
-            .unwrap();
-        tokio::time::sleep(Duration::from_millis(400)).await;
-
-        // Drain the one-shot reap signal.
-        let reply = h.handle.run_batch(a, "SELECT 1".into()).await.unwrap();
-        assert_eq!(error_number(&reply), Some(1205));
-
-        // sp1 belonged to the reaped transaction: rolling back to it in a new
-        // transaction must error 3908, not silently truncate this one's work.
-        let reply = h
-            .handle
-            .run_batch(
-                a,
-                "BEGIN TRAN; INSERT INTO t VALUES (7); INSERT INTO t VALUES (8); ROLLBACK TRANSACTION sp1;"
-                    .into(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3908),
-            "a savepoint from the reaped transaction must not survive"
-        );
-        h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-
-        // And the new transaction's work was never silently discarded.
-        h.handle
-            .run_batch(
-                a,
-                "BEGIN TRAN; INSERT INTO t VALUES (7); INSERT INTO t VALUES (8); COMMIT;".into(),
-            )
-            .await
-            .unwrap();
-        let reply = h
-            .handle
-            .run_batch(a, "SELECT id FROM t ORDER BY id".into())
-            .await
-            .unwrap();
-        assert_eq!(ids(&reply), vec![7, 8], "both committed rows survive");
-    }
-
-    #[tokio::test]
-    async fn writer_blocks_reader_until_commit() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        // A opens a transaction and writes, holding X on t.
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (1);".into())
-            .await
-            .unwrap();
-
-        // B's read must block (READ COMMITTED cannot read A's uncommitted row).
-        let handle_b = h.handle.clone();
-        let read = tokio::spawn(async move {
-            handle_b
-                .run_batch(b, "SELECT id FROM t".into())
-                .await
-                .unwrap()
-        });
-        tokio::time::sleep(Duration::from_millis(150)).await;
-        assert!(
-            !read.is_finished(),
-            "reader should be blocked by the writer"
-        );
-
-        // A commits → releases X → B unblocks and sees the committed row.
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let out = tokio::time::timeout(Duration::from_secs(5), read)
-            .await
-            .expect("reader should unblock after commit")
-            .unwrap();
-        assert_eq!(ids(&out), vec![1]);
-    }
-
-    #[tokio::test]
-    async fn point_writers_to_different_rows_run_concurrently() {
-        // The Stage 12 row-lock win: two transactions updating *different* rows
-        // of one table no longer serialize (Table IX + distinct Row X locks).
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(
-                a,
-                "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT); INSERT INTO t VALUES (1,0),(2,0);".into(),
-            )
-            .await
-            .unwrap();
-
-        // A holds Row X on id = 1 inside an open transaction.
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET v = 10 WHERE id = 1;".into())
-            .await
-            .unwrap();
-
-        // B updates id = 2 — a different row — and must complete without waiting
-        // for A's commit.
-        let out = tokio::time::timeout(
-            Duration::from_secs(3),
-            h.handle
-                .run_batch(b, "UPDATE t SET v = 20 WHERE id = 2".into()),
-        )
-        .await
-        .expect("a point write to a different row must not block")
-        .unwrap();
-        assert!(error_number(&out).is_none(), "{:?}", out.outcome.error);
-
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn point_writers_to_the_same_row_serialize() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(
-                a,
-                "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT); INSERT INTO t VALUES (1,0);"
-                    .into(),
-            )
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET v = 10 WHERE id = 1;".into())
-            .await
-            .unwrap();
-
-        // B updates the *same* row (id = 1): it must block on A's Row X.
-        let handle_b = h.handle.clone();
-        let write = tokio::spawn(async move {
-            handle_b
-                .run_batch(b, "UPDATE t SET v = 20 WHERE id = 1".into())
-                .await
-                .unwrap()
-        });
-        tokio::time::sleep(Duration::from_millis(150)).await;
-        assert!(!write.is_finished(), "same-row writer must block");
-
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let out = tokio::time::timeout(Duration::from_secs(5), write)
-            .await
-            .expect("writer unblocks after commit")
-            .unwrap();
-        assert!(error_number(&out).is_none(), "{:?}", out.outcome.error);
-    }
-
-    #[tokio::test]
-    async fn read_uncommitted_sees_uncommitted_rows_without_blocking() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (7);".into())
-            .await
-            .unwrap();
-
-        // B under READ UNCOMMITTED takes no read lock → dirty-reads A's row.
-        h.handle
-            .run_batch(b, "SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED".into())
-            .await
-            .unwrap();
-        let out = tokio::time::timeout(
-            Duration::from_secs(2),
-            h.handle.run_batch(b, "SELECT id FROM t".into()),
-        )
-        .await
-        .expect("READ UNCOMMITTED must not block")
-        .unwrap();
-        assert_eq!(ids(&out), vec![7], "dirty read sees the uncommitted row");
-
-        h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn disconnect_releases_locks_and_wakes_waiter() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (1);".into())
-            .await
-            .unwrap();
-
-        let handle_b = h.handle.clone();
-        let read = tokio::spawn(async move {
-            handle_b
-                .run_batch(b, "SELECT id FROM t".into())
-                .await
-                .unwrap()
-        });
-        tokio::time::sleep(Duration::from_millis(150)).await;
-        assert!(!read.is_finished(), "reader blocked by open writer txn");
-
-        // A disconnects mid-transaction: rollback releases the lock, waking B,
-        // which now sees an empty table (the insert was undone).
-        h.handle.close_session(a);
-        let out = tokio::time::timeout(Duration::from_secs(5), read)
-            .await
-            .expect("reader should unblock on disconnect")
-            .unwrap();
-        assert_eq!(ids(&out), Vec::<i64>::new());
-    }
-
-    #[tokio::test]
-    async fn deadlock_is_broken_by_timeout_with_1205() {
-        // Wide enough that BOTH conflicting batches park (forming the cycle)
-        // before any deadline expires: since the 1222 split, an expired
-        // waiter with no cycle is reaped as a lock timeout, and a loaded
-        // runner delaying the second park past the deadline would otherwise
-        // turn this test's deadlock into a 1222.
-        let h = start(Duration::from_secs(2));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-
-        for stmt in [
-            "CREATE TABLE a (id INT NOT NULL PRIMARY KEY)",
-            "CREATE TABLE b (id INT NOT NULL PRIMARY KEY)",
-            "INSERT INTO a VALUES (1)",
-            "INSERT INTO b VALUES (1)",
-        ] {
-            h.handle.run_batch(a, stmt.into()).await.unwrap();
-        }
-
-        // A locks table a; B locks table b (each in its own transaction).
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE a SET id = id".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(b, "BEGIN TRAN; UPDATE b SET id = id".into())
-            .await
-            .unwrap();
-
-        // Now each waits for the other's table → deadlock.
-        let ha = h.handle.clone();
-        let a_waits =
-            tokio::spawn(async move { ha.run_batch(a, "UPDATE b SET id = id".into()).await });
-        let hb = h.handle.clone();
-        let b_waits =
-            tokio::spawn(async move { hb.run_batch(b, "UPDATE a SET id = id".into()).await });
-
-        let a_out = tokio::time::timeout(Duration::from_secs(5), a_waits)
-            .await
-            .expect("a_waits resolved")
-            .unwrap()
-            .unwrap();
-        let b_out = tokio::time::timeout(Duration::from_secs(5), b_waits)
-            .await
-            .expect("b_waits resolved")
-            .unwrap()
-            .unwrap();
-
-        // Exactly one is the deadlock victim (1205); the other succeeds.
-        let victims = [&a_out, &b_out]
-            .iter()
-            .filter(|o| error_number(o) == Some(1205))
-            .count();
-        assert_eq!(victims, 1, "exactly one transaction is the deadlock victim");
-    }
-
-    #[tokio::test]
-    async fn deadlock_is_broken_by_the_waits_for_graph_not_the_timeout() {
-        // A 30 s wait timeout: if the deadlock were only broken by the timeout
-        // backstop this would not resolve for 30 s. The waits-for-graph detector
-        // must break it the instant the cycle closes, so the whole thing
-        // finishes well under the timeout.
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        for stmt in [
-            "CREATE TABLE a (id INT NOT NULL PRIMARY KEY)",
-            "CREATE TABLE b (id INT NOT NULL PRIMARY KEY)",
-            "INSERT INTO a VALUES (1)",
-            "INSERT INTO b VALUES (1)",
-        ] {
-            h.handle.run_batch(a, stmt.into()).await.unwrap();
-        }
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE a SET id = id".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(b, "BEGIN TRAN; UPDATE b SET id = id".into())
-            .await
-            .unwrap();
-
-        let ha = h.handle.clone();
-        let a_waits =
-            tokio::spawn(async move { ha.run_batch(a, "UPDATE b SET id = id".into()).await });
-        let hb = h.handle.clone();
-        let b_waits =
-            tokio::spawn(async move { hb.run_batch(b, "UPDATE a SET id = id".into()).await });
-
-        // Both resolve far sooner than the 30 s timeout — proving graph detection.
-        let a_out = tokio::time::timeout(Duration::from_secs(3), a_waits)
-            .await
-            .expect("graph must break the deadlock well under the timeout")
-            .unwrap()
-            .unwrap();
-        let b_out = tokio::time::timeout(Duration::from_secs(3), b_waits)
-            .await
-            .expect("graph must break the deadlock well under the timeout")
-            .unwrap()
-            .unwrap();
-
-        let victims = [&a_out, &b_out]
-            .iter()
-            .filter(|o| error_number(o) == Some(1205))
-            .count();
-        assert_eq!(victims, 1, "exactly one transaction is the deadlock victim");
-    }
-
-    #[tokio::test]
-    async fn deadlock_through_a_fifo_yield_is_detected_by_the_graph() {
-        // A deadlock whose cycle passes through a FIFO anti-barging yield (not a
-        // held-lock conflict): A holds t1; C parks wanting t1+t2; A then wants
-        // the *free* t2 but yields to C, which is queued ahead for it. The graph
-        // must model that yield edge and break the cycle under the 30 s timeout.
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let c = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        for stmt in [
-            "CREATE TABLE t1 (id INT NOT NULL PRIMARY KEY)",
-            "CREATE TABLE t2 (id INT NOT NULL PRIMARY KEY)",
-            "INSERT INTO t1 VALUES (1)",
-            "INSERT INTO t2 VALUES (1)",
-        ] {
-            h.handle.run_batch(a, stmt.into()).await.unwrap();
-        }
-        // A holds X(t1).
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t1 SET id = id".into())
-            .await
-            .unwrap();
-        // C wants t2 then t1 (held by A) → parks, now queued ahead for t2.
-        let hc = h.handle.clone();
-        let c_waits = tokio::spawn(async move {
-            hc.run_batch(
-                c,
-                "BEGIN TRAN; UPDATE t2 SET id = id; UPDATE t1 SET id = id".into(),
-            )
-            .await
-        });
-        // Ensure C is parked before A asks for t2, so A queues behind it.
-        tokio::time::sleep(Duration::from_millis(250)).await;
-        // A wants the free t2 but yields to C (ahead) → parks → FIFO cycle.
-        let ha = h.handle.clone();
-        let a_waits =
-            tokio::spawn(async move { ha.run_batch(a, "UPDATE t2 SET id = id".into()).await });
-
-        let a_out = tokio::time::timeout(Duration::from_secs(3), a_waits)
-            .await
-            .expect("graph must break the FIFO deadlock well under the timeout")
-            .unwrap()
-            .unwrap();
-        let c_out = tokio::time::timeout(Duration::from_secs(3), c_waits)
-            .await
-            .expect("graph must break the FIFO deadlock well under the timeout")
-            .unwrap()
-            .unwrap();
-        let victims = [&a_out, &c_out]
-            .iter()
-            .filter(|o| error_number(o) == Some(1205))
-            .count();
-        assert_eq!(victims, 1, "exactly one transaction is the deadlock victim");
-    }
-
-    #[tokio::test]
-    async fn repeatable_read_holds_shared_lock_and_blocks_a_writer() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "INSERT INTO t VALUES (1)".into())
-            .await
-            .unwrap();
-
-        // A reads under REPEATABLE READ inside a transaction → holds S on t.
-        h.handle
-            .run_batch(a, "SET TRANSACTION ISOLATION LEVEL REPEATABLE READ".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; SELECT id FROM t;".into())
-            .await
-            .unwrap();
-
-        // B's write must block on A's retained shared lock.
-        let hb = h.handle.clone();
-        let write = tokio::spawn(async move {
-            hb.run_batch(b, "UPDATE t SET id = id".into())
-                .await
-                .unwrap()
-        });
-        tokio::time::sleep(Duration::from_millis(150)).await;
-        assert!(
-            !write.is_finished(),
-            "REPEATABLE READ keeps the shared lock, blocking the writer"
-        );
-
-        // A commits → releases S → B proceeds.
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let out = tokio::time::timeout(Duration::from_secs(5), write)
-            .await
-            .expect("writer unblocks after reader commits")
-            .unwrap();
-        assert!(out.outcome.error.is_none(), "{:?}", out.outcome.error);
-    }
-
-    #[tokio::test]
-    async fn read_committed_releases_shared_lock_so_a_writer_proceeds() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "INSERT INTO t VALUES (1)".into())
-            .await
-            .unwrap();
-
-        // A reads under READ COMMITTED (the default) inside a transaction; its
-        // shared lock is dropped at statement end even though the txn stays open.
-        h.handle
-            .run_batch(a, "BEGIN TRAN; SELECT id FROM t;".into())
-            .await
-            .unwrap();
-
-        // B's write is NOT blocked — unlike REPEATABLE READ above.
-        let out = tokio::time::timeout(
-            Duration::from_secs(2),
-            h.handle.run_batch(b, "UPDATE t SET id = id".into()),
-        )
-        .await
-        .expect("READ COMMITTED releases the shared lock, so the writer runs")
-        .unwrap();
-        assert!(out.outcome.error.is_none(), "{:?}", out.outcome.error);
-
-        h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn isolation_escalation_within_a_batch_locks_the_read() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "INSERT INTO t VALUES (1)".into())
-            .await
-            .unwrap();
-
-        // A holds X on t.
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET id = id;".into())
-            .await
-            .unwrap();
-
-        // B is READ UNCOMMITTED, so a plain read would take no lock and could
-        // dirty-read. But B raises the level to SERIALIZABLE in the SAME batch
-        // as the read, which must lock the read → it blocks on A's writer.
-        h.handle
-            .run_batch(b, "SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED".into())
-            .await
-            .unwrap();
-        let hb = h.handle.clone();
-        let read = tokio::spawn(async move {
-            hb.run_batch(
-                b,
-                "SET TRANSACTION ISOLATION LEVEL SERIALIZABLE; SELECT id FROM t;".into(),
-            )
-            .await
-            .unwrap()
-        });
-        tokio::time::sleep(Duration::from_millis(150)).await;
-        assert!(
-            !read.is_finished(),
-            "an escalated read must lock and block on the uncommitted writer"
-        );
-
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let out = tokio::time::timeout(Duration::from_secs(5), read)
-            .await
-            .expect("escalated read unblocks after commit")
-            .unwrap();
-        assert_eq!(ids(&out), vec![1]);
-    }
-
-    #[tokio::test]
-    async fn holder_is_not_blocked_by_a_waiter_on_its_own_lock() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "INSERT INTO t VALUES (1)".into())
-            .await
-            .unwrap();
-
-        // A holds X on t via an open transaction.
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET id = id;".into())
-            .await
-            .unwrap();
-
-        // B blocks on t and parks in the queue.
-        let hb = h.handle.clone();
-        let b_read =
-            tokio::spawn(async move { hb.run_batch(b, "SELECT id FROM t".into()).await.unwrap() });
-        tokio::time::sleep(Duration::from_millis(150)).await;
-        assert!(!b_read.is_finished(), "B parks on A's lock");
-
-        // A re-touches t in a new batch. Because A already holds the lock, it
-        // must NOT yield to the queued waiter B — doing so would deadlock A on
-        // its own lock. This completes promptly.
-        let again = tokio::time::timeout(
-            Duration::from_secs(2),
-            h.handle.run_batch(a, "UPDATE t SET id = id".into()),
-        )
-        .await
-        .expect("holder must not self-deadlock on a waiter behind its own lock")
-        .unwrap();
-        assert!(again.outcome.error.is_none(), "{:?}", again.outcome.error);
-
-        // A commits → B finally proceeds.
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let out = tokio::time::timeout(Duration::from_secs(5), b_read)
-            .await
-            .expect("B unblocks after A commits")
-            .unwrap();
-        assert_eq!(ids(&out), vec![1]);
-    }
-
-    #[tokio::test]
-    async fn autocommit_reads_run_concurrently() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "INSERT INTO t VALUES (1)".into())
-            .await
-            .unwrap();
-
-        // Two shared readers never block each other.
-        let out_a = h
-            .handle
-            .run_batch(a, "SELECT id FROM t".into())
-            .await
-            .unwrap();
-        let out_b = tokio::time::timeout(
-            Duration::from_secs(2),
-            h.handle.run_batch(b, "SELECT id FROM t".into()),
-        )
-        .await
-        .expect("concurrent shared reads must not block")
-        .unwrap();
-        assert_eq!(ids(&out_a), vec![1]);
-        assert_eq!(ids(&out_b), vec![1]);
-    }
-
-    /// The worker pool's core correctness stress test: many sessions run money
-    /// transfers concurrently, some rolled back at random, and the total across
-    /// all accounts must be exactly conserved — no lost updates, no torn
-    /// transactions, no money created or destroyed by the concurrent plumbing
-    /// (take/return of session context, parking, waking, draining).
-    #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-    async fn concurrent_transfers_conserve_the_total() {
-        const ACCOUNTS: i64 = 8;
-        const TASKS: usize = 16;
-        const TRANSFERS: usize = 25;
-        const INITIAL: i64 = 1000;
-
-        let h = start(Duration::from_secs(30));
-        let setup = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(
-                setup,
-                "CREATE TABLE accounts (id INT NOT NULL PRIMARY KEY, balance INT NOT NULL)".into(),
-            )
-            .await
-            .unwrap();
-        for id in 1..=ACCOUNTS {
-            h.handle
-                .run_batch(
-                    setup,
-                    format!("INSERT INTO accounts VALUES ({id}, {INITIAL})"),
-                )
-                .await
-                .unwrap();
-        }
-
-        let mut tasks = Vec::new();
-        for t in 0..TASKS {
-            let handle = h.handle.clone();
-            tasks.push(tokio::spawn(async move {
-                let session = handle
-                    .open_session(String::new(), String::new(), 0)
-                    .await
-                    .expect("open session")
-                    .0;
-                // A deterministic per-task PRNG (an LCG) — reproducible, no dep.
-                let mut rng: u64 =
-                    0x9E37_79B9_7F4A_7C15 ^ (t as u64).wrapping_mul(0x2545_F491_4F6C_DD1D);
-                let mut next = move || {
-                    rng = rng
-                        .wrapping_mul(6364136223846793005)
-                        .wrapping_add(1442695040888963407);
-                    rng >> 33
-                };
-                for _ in 0..TRANSFERS {
-                    let a = (next() % ACCOUNTS as u64) as i64 + 1;
-                    // Force a distinct counterparty: (a % N) + 1 is never a.
-                    let b = (a % ACCOUNTS) + 1;
-                    let amount = (next() % 50) as i64 + 1;
-                    // One in ten transactions rolls back; conservation must hold
-                    // either way.
-                    let close = if next() % 10 == 0 {
-                        "ROLLBACK"
-                    } else {
-                        "COMMIT"
-                    };
-                    let sql = format!(
-                        "BEGIN TRAN; \
-                         UPDATE accounts SET balance = balance - {amount} WHERE id = {a}; \
-                         UPDATE accounts SET balance = balance + {amount} WHERE id = {b}; \
-                         {close};"
-                    );
-                    // A deadlock victim (1205) rolls back cleanly; just retry it.
-                    loop {
-                        let reply = handle.run_batch(session, sql.clone()).await.unwrap();
-                        match error_number(&reply) {
-                            Some(1205) => continue,
-                            Some(other) => panic!("unexpected error {other} on transfer"),
-                            None => break,
-                        }
-                    }
-                }
-                handle.close_session(session);
-            }));
-        }
-        for task in tasks {
-            task.await.unwrap();
-        }
-
-        let reply = h
-            .handle
-            .run_batch(setup, "SELECT balance FROM accounts".into())
-            .await
-            .unwrap();
-        let total: i64 = ids(&reply).iter().sum();
-        assert_eq!(
-            total,
-            ACCOUNTS * INITIAL,
-            "money was created or destroyed under concurrency"
-        );
-    }
-
-    /// Like the conservation test, but each transfer is a *multi-batch*
-    /// transaction across two tables, so locks are taken incrementally and two
-    /// transfers in opposite table order can genuinely deadlock. A 1205 victim
-    /// rolls back and retries. Exercises cross-batch lock holding, the deadlock
-    /// detector, and victim retry all under concurrent load — the total must
-    /// still be conserved.
-    #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-    async fn concurrent_multi_table_transfers_survive_deadlocks() {
-        const ACCOUNTS: i64 = 3;
-        const TASKS: usize = 10;
-        const TRANSFERS: usize = 12;
-        const INITIAL: i64 = 1000;
-        const TABLES: [&str; 2] = ["checking", "savings"];
-
-        let h = start(Duration::from_secs(30));
-        let setup = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        for table in TABLES {
-            h.handle
-                .run_batch(
-                    setup,
-                    format!(
-                        "CREATE TABLE {table} (id INT NOT NULL PRIMARY KEY, balance INT NOT NULL)"
-                    ),
-                )
-                .await
-                .unwrap();
-            for id in 1..=ACCOUNTS {
-                h.handle
-                    .run_batch(
-                        setup,
-                        format!("INSERT INTO {table} VALUES ({id}, {INITIAL})"),
-                    )
-                    .await
-                    .unwrap();
-            }
-        }
-
-        let mut tasks = Vec::new();
-        for t in 0..TASKS {
-            let handle = h.handle.clone();
-            tasks.push(tokio::spawn(async move {
-                let session = handle
-                    .open_session(String::new(), String::new(), 0)
-                    .await
-                    .expect("open session")
-                    .0;
-                let mut rng: u64 =
-                    0xDEAD_BEEF_0000_0001 ^ (t as u64).wrapping_mul(0x2545_F491_4F6C_DD1D);
-                let mut next = move || {
-                    rng = rng
-                        .wrapping_mul(6364136223846793005)
-                        .wrapping_add(1442695040888963407);
-                    rng >> 33
-                };
-                for _ in 0..TRANSFERS {
-                    let src = TABLES[(next() % 2) as usize];
-                    let dst = TABLES[(next() % 2) as usize];
-                    let a = (next() % ACCOUNTS as u64) as i64 + 1;
-                    let b = (next() % ACCOUNTS as u64) as i64 + 1;
-                    let amount = (next() % 40) as i64 + 1;
-                    let close = if next() % 8 == 0 {
-                        "ROLLBACK"
-                    } else {
-                        "COMMIT"
-                    };
-                    // Deadlock victims only ever land on the two UPDATEs; BEGIN
-                    // and COMMIT/ROLLBACK take no new locks. Retry the whole
-                    // transaction (already rolled back) from the top.
-                    'attempt: loop {
-                        let steps = [
-                            "BEGIN TRAN".to_string(),
-                            format!("UPDATE {src} SET balance = balance - {amount} WHERE id = {a}"),
-                            format!("UPDATE {dst} SET balance = balance + {amount} WHERE id = {b}"),
-                            close.to_string(),
-                        ];
-                        for step in steps {
-                            let reply = handle.run_batch(session, step).await.unwrap();
-                            match error_number(&reply) {
-                                Some(1205) => continue 'attempt,
-                                Some(other) => panic!("unexpected error {other}"),
-                                None => {}
-                            }
-                        }
-                        break;
-                    }
-                }
-                handle.close_session(session);
-            }));
-        }
-        for task in tasks {
-            task.await.unwrap();
-        }
-
-        let mut total = 0i64;
-        for table in TABLES {
-            let reply = h
-                .handle
-                .run_batch(setup, format!("SELECT balance FROM {table}"))
-                .await
-                .unwrap();
-            total += ids(&reply).iter().sum::<i64>();
-        }
-        assert_eq!(
-            total,
-            2 * ACCOUNTS * INITIAL,
-            "money not conserved across tables under deadlock retries"
-        );
-    }
-
-    /// A single waiter blocked on a legitimately-held lock (no cycle, so the
-    /// graph detector finds nothing) must still be freed by the lock-wait
-    /// timeout even when the pool then goes completely quiet — no further call
-    /// arrives to wake a worker. Regression: a stale `earliest_deadline`
-    /// snapshot used to let a worker block in an untimed `recv` holding the
-    /// single-consumer rx mutex, disabling the reaper during quiescence and
-    /// hanging the waiter indefinitely instead of timing it out.
-    #[tokio::test]
-    async fn lone_waiter_is_reaped_by_timeout_when_pool_goes_idle() {
-        let h = start(Duration::from_millis(300));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-
-        h.handle
-            .run_batch(a, "CREATE TABLE t (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "INSERT INTO t VALUES (1)".into())
-            .await
-            .unwrap();
-        // A holds X on t and stays idle inside its transaction (never commits).
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET id = id".into())
-            .await
-            .unwrap();
-
-        // B's read conflicts with A's uncommitted write and parks. There is no
-        // cycle (A waits on nothing), so only the timeout backstop can free it,
-        // and no further calls arrive to wake a worker.
-        let out = tokio::time::timeout(
-            Duration::from_secs(5),
-            h.handle.run_batch(b, "SELECT id FROM t".into()),
-        )
-        .await
-        .expect("lone waiter must be reaped by the timeout, not hang forever")
-        .unwrap();
-        assert_eq!(
-            error_number(&out),
-            Some(1222),
-            "a lone waiter behind a live holder times out as 1222 — there is \
-             no cycle, so 1205 would report a deadlock that never happened"
-        );
-    }
-
-    /// Stage 12's exit case: the engine stays responsive during a large
-    /// scan. A protocol heartbeat never touches the engine (the dispatcher
-    /// answers it), so the meaningful half is a native search completing
-    /// WHILE a worker is mid-scan holding the batch-long read gate — pinned
-    /// by synchronizing on the scan's first Columns event (the worker is
-    /// provably inside the batch) and asserting the search returns before
-    /// the scan finishes. A regression to an exclusive gate blocks the
-    /// search behind the whole scan.
-    #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-    async fn a_search_is_answered_while_a_large_scan_runs() {
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        fill(&h, s, 60_000).await;
-        h.handle
-            .run_native(
-                r#"create index bench { "mappings": { "properties": { "title": { "type": "text" } } } }"#
-                    .to_string(),
-            )
-            .await
-            .expect("create index");
-        h.handle
-            .run_native(r#"insert document bench { "title": "hello world" }"#.to_string())
-            .await
-            .expect("insert doc");
-
-        let mut events = h
-            .handle
-            .stream_batch(s, "SELECT id FROM t".into(), no_cancel());
-        // The first Columns event: the worker is inside the batch, gate held.
-        loop {
-            match events.recv().await.expect("stream lives") {
-                BatchEvent::Columns(_) => break,
-                BatchEvent::Failed(err) => panic!("scan failed to start: {err}"),
-                _ => {}
-            }
-        }
-
-        let search = h
-            .handle
-            .run_native(r#"search bench {"query": {"match": {"title": "hello"}}}"#.to_string())
-            .await
-            .expect("search");
-        assert!(search.contains("hello world"), "search answered: {search}");
-        // The scan must have STILL BEEN RUNNING when the search returned.
-        // The channel is unbounded, so "the next event" proves nothing —
-        // instead drain everything already buffered at this instant without
-        // blocking: if the batch's terminal event is among it, the batch
-        // finished before the search did (an exclusive gate makes the search
-        // wait out the whole scan, and this assertion catches exactly that).
-        let mut already_finished = false;
-        while let Ok(event) = events.try_recv() {
-            if matches!(event, BatchEvent::Complete { .. } | BatchEvent::Failed(_)) {
-                already_finished = true;
-            }
-        }
-        assert!(
-            !already_finished,
-            "the search must complete while the scan is mid-stream, not after it"
-        );
-        drain_events(events).await;
-    }
-
-    // ---- sp_prepare handle family -----------------------------------------
-
-    /// The PreparedHandle event's value, if the stream carried one.
-    fn handle_of(events: &[BatchEvent]) -> Option<i32> {
-        events.iter().find_map(|event| match event {
-            BatchEvent::PreparedHandle(h) => Some(*h),
-            _ => None,
-        })
-    }
-
-    /// The first Error event's number, if any.
-    fn event_error(events: &[BatchEvent]) -> Option<i32> {
-        events.iter().find_map(|event| match event {
-            BatchEvent::Error(e) => Some(e.number),
-            _ => None,
-        })
-    }
-
-    fn int_param(value: i32) -> crate::engine::RpcParam {
-        crate::engine::RpcParam {
-            name: String::new(),
-            column_type: crate::relstore::types::ColumnType::Int,
-            value: Datum::Int(value),
-        }
-    }
-
-    /// All streamed rows, flattened.
-    fn rows_of(events: &[BatchEvent]) -> Vec<Vec<Datum>> {
-        events
-            .iter()
-            .filter_map(|event| match event {
-                BatchEvent::Rows(rows) => Some(rows.clone()),
-                _ => None,
-            })
-            .flatten()
-            .collect()
-    }
-
-    #[tokio::test]
-    async fn prepare_execute_unprepare_roundtrip() {
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        fill(&h, s, 5).await;
-
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Prepare {
-                decls: "@p1 int".into(),
-                stmt: "SELECT id FROM t WHERE id = @p1".into(),
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(event_error(&events), None, "{events:?}");
-        let handle = handle_of(&events).expect("prepare returns a handle");
-
-        // Execute twice with different values: the same handle re-binds each
-        // time, and the unnamed wire value takes the declaration's name.
-        for wanted in [3, 5] {
-            let events = drain_events(h.handle.stream_prepared(
-                s,
-                PreparedRpc::Execute {
-                    handle,
-                    values: vec![int_param(wanted)],
-                },
-                no_cancel(),
-            ))
-            .await;
-            assert_eq!(event_error(&events), None, "{events:?}");
-            assert_eq!(rows_of(&events), vec![vec![Datum::Int(wanted)]]);
-        }
-
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Unprepare { handle },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(event_error(&events), None, "{events:?}");
-
-        // The dropped handle is gone: 8179, SQL Server's number for it.
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Execute {
-                handle,
-                values: vec![int_param(1)],
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(event_error(&events), Some(8179), "{events:?}");
-    }
-
-    #[tokio::test]
-    async fn an_unknown_handle_answers_8179() {
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Execute {
-                handle: 42,
-                values: Vec::new(),
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(event_error(&events), Some(8179));
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Unprepare { handle: 42 },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(event_error(&events), Some(8179));
-    }
-
-    #[tokio::test]
-    async fn a_parse_error_at_prepare_allocates_no_handle() {
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Prepare {
-                decls: String::new(),
-                stmt: "SELEC oops".into(),
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert!(event_error(&events).is_some(), "{events:?}");
-        assert_eq!(handle_of(&events), None, "no handle on a failed prepare");
-
-        // The failed prepare consumed nothing: the next handle is still 1.
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Prepare {
-                decls: String::new(),
-                stmt: "SELECT 1 AS one".into(),
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(handle_of(&events), Some(1));
-    }
-
-    #[tokio::test]
-    async fn ddl_between_prepare_and_execute_sees_the_new_schema() {
-        // There is no cached plan: every execute re-binds against the live
-        // catalog, so DDL between prepare and execute needs no invalidation —
-        // the same handle simply sees the new schema (the plan's
-        // catalog_version/rebind machinery is moot by design).
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        fill(&h, s, 3).await;
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Prepare {
-                decls: String::new(),
-                stmt: "SELECT * FROM t ORDER BY id".into(),
-            },
-            no_cancel(),
-        ))
-        .await;
-        let handle = handle_of(&events).expect("prepare returns a handle");
-
-        h.handle
-            .run_batch(
-                s,
-                "DROP TABLE t; CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT NOT NULL); INSERT INTO t VALUES (7, 70)".into(),
-            )
-            .await
-            .unwrap();
-
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Execute {
-                handle,
-                values: Vec::new(),
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(event_error(&events), None, "{events:?}");
-        // The wildcard expands against the NEW table: two columns, new row.
-        assert_eq!(rows_of(&events), vec![vec![Datum::Int(7), Datum::Int(70)]]);
-    }
-
-    #[tokio::test]
-    async fn prepexec_reports_the_handle_after_the_results() {
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        fill(&h, s, 4).await;
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::PrepExec {
-                decls: "@p1 int".into(),
-                stmt: "SELECT id FROM t WHERE id > @p1 ORDER BY id".into(),
-                values: vec![int_param(2)],
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(event_error(&events), None, "{events:?}");
-        assert_eq!(
-            rows_of(&events),
-            vec![vec![Datum::Int(3)], vec![Datum::Int(4)]]
-        );
-        // Return values follow every result set: the handle event comes after
-        // the statement's DONE, immediately before Complete.
-        let positions: Vec<&str> = events
-            .iter()
-            .map(|e| match e {
-                BatchEvent::Columns(_) => "columns",
-                BatchEvent::Rows(_) => "rows",
-                BatchEvent::StatementDone { .. } => "done",
-                BatchEvent::PreparedHandle(_) => "handle",
-                BatchEvent::Complete { .. } => "complete",
-                _ => "other",
-            })
-            .collect();
-        assert_eq!(
-            positions.iter().rev().take(2).copied().collect::<Vec<_>>(),
-            ["complete", "handle"],
-            "{positions:?}"
-        );
-        // And the stored handle is executable afterwards.
-        let handle = handle_of(&events).expect("prepexec returns a handle");
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Execute {
-                handle,
-                values: vec![int_param(3)],
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(rows_of(&events), vec![vec![Datum::Int(4)]]);
-    }
-
-    #[tokio::test]
-    async fn describe_first_result_set_covers_static_shapes_only() {
-        let h = start(LOCK_WAIT_TIMEOUT);
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(
-                s,
-                "CREATE TABLE d (id INT NOT NULL PRIMARY KEY, name NVARCHAR(40))".into(),
-            )
-            .await
-            .unwrap();
-
-        // A parameterized single-table SELECT describes without executing.
-        // The @p1 is unresolvable in describe's default context — the planner
-        // swallows that and falls back to a scan (non-sargable predicate);
-        // the columns come from the table schema either way.
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Describe {
-                tsql: "SELECT id, name FROM d WHERE id = @p1".into(),
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(event_error(&events), None, "{events:?}");
-        let rows = rows_of(&events);
-        assert_eq!(rows.len(), 2);
-        assert_eq!(rows[0][1], Datum::Int(1)); // column_ordinal
-        assert_eq!(rows[0][2], Datum::NVarChar("id".into()));
-        assert_eq!(rows[0][4], Datum::Int(56)); // system_type_id: int
-        assert_eq!(rows[1][2], Datum::NVarChar("name".into()));
-        assert_eq!(rows[1][5], Datum::NVarChar("nvarchar(40)".into()));
-
-        // A statement producing no result set describes as zero rows — and
-        // describing it executes nothing: the table stays empty.
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Describe {
-                tsql: "INSERT INTO d VALUES (1, 'x')".into(),
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(event_error(&events), None, "{events:?}");
-        assert_eq!(rows_of(&events), Vec::<Vec<Datum>>::new());
-        let reply = h
-            .handle
-            .run_batch(s, "SELECT COUNT(*) FROM d".into())
-            .await
-            .unwrap();
-        match &reply.outcome.results[0] {
-            StatementResult::Rows(rowset) => assert_eq!(
-                rowset.rows[0][0],
-                Datum::BigInt(0),
-                "describe must not execute the INSERT"
-            ),
-            other => panic!("expected rows, got {other:?}"),
-        }
-
-        // The contract is the first RESULT SET, not the first statement:
-        // `INSERT; SELECT` (and a SELECT inside a TRY block) describe the
-        // SELECT, and `TOP 0` — the standard metadata-discovery idiom —
-        // describes like any other TOP.
-        for tsql in [
-            "INSERT INTO d VALUES (9, 'y'); SELECT id FROM d",
-            "BEGIN TRY SELECT id FROM d END TRY BEGIN CATCH END CATCH",
-            "SELECT TOP 0 id FROM d",
-        ] {
-            let events = drain_events(h.handle.stream_prepared(
-                s,
-                PreparedRpc::Describe { tsql: tsql.into() },
-                no_cancel(),
-            ))
-            .await;
-            assert_eq!(event_error(&events), None, "{tsql}: {events:?}");
-            let rows = rows_of(&events);
-            assert_eq!(rows.len(), 1, "{tsql}");
-            assert_eq!(rows[0][2], Datum::NVarChar("id".into()), "{tsql}");
-        }
-
-        // A shape whose types are only known by executing answers 11514.
-        let events = drain_events(h.handle.stream_prepared(
-            s,
-            PreparedRpc::Describe {
-                tsql: "SELECT a.id FROM d a JOIN d b ON a.id = b.id".into(),
-            },
-            no_cancel(),
-        ))
-        .await;
-        assert_eq!(event_error(&events), Some(11514), "{events:?}");
-    }
-
-    #[test]
-    fn decl_names_splits_top_level_commas_only() {
-        assert_eq!(
-            crate::engine::decl_names("@p1 int, @p2 nvarchar(10), @p3 decimal(10,2)"),
-            ["@p1", "@p2", "@p3"]
-        );
-        assert_eq!(crate::engine::decl_names(""), Vec::<String>::new());
-        assert_eq!(crate::engine::decl_names("@a int"), ["@a"]);
-        // A quoted default may contain commas and parens; a doubled ''
-        // escape stays inside the string.
-        assert_eq!(
-            crate::engine::decl_names(
-                "@p1 varchar(10) = 'a,b', @p2 int, @p3 varchar(5) = 'it''s, ok'"
-            ),
-            ["@p1", "@p2", "@p3"]
-        );
-    }
-
-    #[test]
-    fn bind_decl_names_keeps_existing_names() {
-        let mut named = int_param(1);
-        named.name = "@mine".into();
-        let bound = bind_decl_names("@p1 int, @p2 int", vec![named, int_param(2)]).expect("bind");
-        assert_eq!(bound[0].name, "@mine");
-        assert_eq!(bound[1].name, "@p2");
-    }
-
-    #[test]
-    fn more_values_than_declarations_is_8144() {
-        // SQL Server rejects extra arguments rather than silently ignoring
-        // them; without this an unmatched value seeded nothing and vanished.
-        let err = bind_decl_names("@p1 int", vec![int_param(1), int_param(2)])
-            .expect_err("extra value must be rejected");
-        assert_eq!(err.number, 8144);
-        let err = bind_decl_names("", vec![int_param(1)])
-            .expect_err("values against an empty declaration list must be rejected");
-        assert_eq!(err.number, 8144);
-        // Fewer values than declarations stays legal (an unread declared
-        // parameter goes unmissed).
-        assert!(bind_decl_names("@p1 int, @p2 int", vec![int_param(1)]).is_ok());
-        // Extra NAMED values pass through — they seed variables by their own
-        // names (the run_rpc wrappers' contract).
-        let mut named = int_param(9);
-        named.name = "@extra".into();
-        assert!(bind_decl_names("", vec![named]).is_ok());
-    }
-
-    // ---- READ_COMMITTED_SNAPSHOT (Stage 13) ------------------------------
-
-    /// A harness with RCSI on and `t (id PK, v)` seeded with (1,10), (2,20).
-    async fn rcsi_harness() -> (Harness, SessionId, SessionId) {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        for sql in [
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            "INSERT INTO t VALUES (1, 10), (2, 20)",
-        ] {
-            let reply = h.handle.run_batch(a, sql.into()).await.unwrap();
-            assert_eq!(
-                error_number(&reply),
-                None,
-                "{sql}: {:?}",
-                reply.outcome.error
-            );
-        }
-        (h, a, b)
-    }
-
-    /// Runs a batch under a 2 s timeout: an RCSI reader must never park
-    /// behind a writer, so a hang here is the failure being tested for.
-    async fn must_not_block(h: &Harness, s: SessionId, sql: &str) -> BatchReply {
-        tokio::time::timeout(Duration::from_secs(2), h.handle.run_batch(s, sql.into()))
-            .await
-            .expect("an RCSI read must not block on a writer")
-            .unwrap()
-    }
-
-    /// Column 0 of the first rowset as i64s, asserting the batch succeeded.
-    fn values(reply: &BatchReply) -> Vec<i64> {
-        assert_eq!(error_number(reply), None, "{:?}", reply.outcome.error);
-        ids(reply)
-    }
-
-    #[tokio::test]
-    async fn rcsi_select_reads_committed_state_without_waiting_for_the_writer() {
-        let (h, a, b) = rcsi_harness().await;
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET v = 99 WHERE id = 1;".into())
-            .await
-            .unwrap();
-        // B reads the pre-update value while A's transaction holds its X
-        // lock: the value is the proof there was no wait (a lock-based
-        // reader could only return 99, after A commits).
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![99]);
-    }
-
-    #[tokio::test]
-    async fn without_rcsi_the_same_reader_waits_for_the_commit() {
-        // The control for the test above — and the teeth of the
-        // `rcsi_enabled` gate: with the option off, the reader parks on the
-        // writer's lock and can only ever see the committed value.
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(
-                a,
-                "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT); \
-                 INSERT INTO t VALUES (1, 10);"
-                    .into(),
-            )
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET v = 99 WHERE id = 1;".into())
-            .await
-            .unwrap();
-        let reader = {
-            let handle = h.handle.clone();
-            tokio::spawn(async move { handle.run_batch(b, "SELECT v FROM t".into()).await })
-        };
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = reader.await.unwrap().unwrap();
-        assert_eq!(
-            values(&reply),
-            vec![99],
-            "a lock-based reader sees only the commit"
-        );
-    }
-
-    #[tokio::test]
-    async fn rcsi_reader_sees_a_deleted_row_until_the_delete_commits() {
-        let (h, a, b) = rcsi_harness().await;
-        h.handle
-            .run_batch(a, "BEGIN TRAN; DELETE FROM t WHERE id = 1;".into())
-            .await
-            .unwrap();
-        // The row is physically gone from the page; the scan's merge pass
-        // serves it from its version image.
-        let reply = must_not_block(&h, b, "SELECT id FROM t").await;
-        let mut got = values(&reply);
-        got.sort();
-        assert_eq!(got, vec![1, 2]);
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT id FROM t").await;
-        assert_eq!(values(&reply), vec![2]);
-    }
-
-    #[tokio::test]
-    async fn rcsi_insert_is_invisible_until_it_commits() {
-        let (h, a, b) = rcsi_harness().await;
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (3, 30);".into())
-            .await
-            .unwrap();
-        let reply = must_not_block(&h, b, "SELECT id FROM t").await;
-        let mut got = values(&reply);
-        got.sort();
-        assert_eq!(
-            got,
-            vec![1, 2],
-            "an uncommitted insert does not exist for the snapshot"
-        );
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT id FROM t").await;
-        let mut got = values(&reply);
-        got.sort();
-        assert_eq!(got, vec![1, 2, 3]);
-    }
-
-    #[tokio::test]
-    async fn rcsi_rekey_update_shows_the_old_row_until_commit() {
-        let (h, a, b) = rcsi_harness().await;
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET id = 9 WHERE id = 1;".into())
-            .await
-            .unwrap();
-        let reply = must_not_block(&h, b, "SELECT id FROM t").await;
-        let mut got = values(&reply);
-        got.sort();
-        assert_eq!(
-            got,
-            vec![1, 2],
-            "the re-key is invisible: old identity present, new absent"
-        );
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT id FROM t").await;
-        let mut got = values(&reply);
-        got.sort();
-        assert_eq!(got, vec![2, 9]);
-    }
-
-    #[tokio::test]
-    async fn rcsi_index_seek_serves_the_snapshot_row() {
-        let (h, a, b) = rcsi_harness().await;
-        // Enough rows that the planner takes the seek (tiny tables demote to
-        // scans), plus an index on v.
-        let mut fill = String::from("INSERT INTO t VALUES ");
-        for i in 3..=25 {
-            fill.push_str(&format!("({}, {}),", i, i * 100));
-        }
-        fill.pop();
-        h.handle.run_batch(a, fill).await.unwrap();
-        h.handle
-            .run_batch(a, "CREATE INDEX ix_v ON t (v)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET v = 4242 WHERE id = 1;".into())
-            .await
-            .unwrap();
-        // The uncommitted update moved the index entry from 10 to 4242. The
-        // snapshot must still find id=1 under v=10 (its entry is gone — the
-        // merge pass restores it) and must not find it under v=4242 (the
-        // entry exists, but the visible image fails the predicate).
-        let reply = must_not_block(&h, b, "SELECT id FROM t WHERE v = 10").await;
-        assert_eq!(values(&reply), vec![1]);
-        let reply = must_not_block(&h, b, "SELECT id FROM t WHERE v = 4242").await;
-        assert_eq!(values(&reply), Vec::<i64>::new());
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT id FROM t WHERE v = 4242").await;
-        assert_eq!(values(&reply), vec![1]);
-        let reply = must_not_block(&h, b, "SELECT id FROM t WHERE v = 10").await;
-        assert_eq!(values(&reply), Vec::<i64>::new());
-    }
-
-    #[tokio::test]
-    async fn rcsi_reader_sees_its_own_uncommitted_writes() {
-        let (h, a, _b) = rcsi_harness().await;
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET v = 55 WHERE id = 1;".into())
-            .await
-            .unwrap();
-        let reply = must_not_block(&h, a, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(
-            values(&reply),
-            vec![55],
-            "a session always sees its own writes"
-        );
-        h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-        let reply = must_not_block(&h, a, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-    }
-
-    #[tokio::test]
-    async fn rcsi_same_batch_write_then_read_sees_the_write() {
-        // The durable-prefix rule must not hide the session's own
-        // just-committed statement from a later statement in the same batch
-        // (the flush-before-capture path).
-        let (h, a, _b) = rcsi_harness().await;
-        let reply = must_not_block(
-            &h,
-            a,
-            "UPDATE t SET v = 11 WHERE id = 1; SELECT v FROM t WHERE id = 1",
-        )
-        .await;
-        assert_eq!(values(&reply), vec![11]);
-    }
-
-    #[tokio::test]
-    async fn rollback_after_publish_keeps_chains_sound_for_reuse() {
-        // A rolled-back insert must unpublish its version entries: the next
-        // insert of the same key builds a fresh chain (the debug asserts on
-        // chain-head invariants fire here if it does not), and readers see
-        // only the committed state.
-        let (h, a, b) = rcsi_harness().await;
-        h.handle
-            .run_batch(a, "BEGIN TRAN; INSERT INTO t VALUES (5, 50);".into())
-            .await
-            .unwrap();
-        h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-        let reply = h
-            .handle
-            .run_batch(a, "INSERT INTO t VALUES (5, 55)".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 5").await;
-        assert_eq!(values(&reply), vec![55]);
-    }
-
-    #[tokio::test]
-    async fn savepoint_rollback_unpublishes_the_rolled_back_statements() {
-        let (h, a, b) = rcsi_harness().await;
-        h.handle
-            .run_batch(
-                a,
-                "BEGIN TRAN; UPDATE t SET v = 100 WHERE id = 1; \
-                 SAVE TRANSACTION s; UPDATE t SET v = 200 WHERE id = 2; \
-                 ROLLBACK TRANSACTION s;"
-                    .into(),
-            )
-            .await
-            .unwrap();
-        // While A is open, B sees both originals (100 and 200 are invisible
-        // either way — the discriminator is after the commit).
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 2").await;
-        assert_eq!(values(&reply), vec![20]);
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![100], "the kept update committed");
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 2").await;
-        assert_eq!(
-            values(&reply),
-            vec![20],
-            "the savepoint-rolled-back update did not"
-        );
-    }
-
-    #[tokio::test]
-    async fn rcsi_assignment_select_sees_the_same_batch_commit() {
-        // An assignment SELECT opens no rowset, so it skips `run_block`'s
-        // pre-rowset durability flush — the snapshot capture must flush on
-        // its own or the durable-prefix rule hides the batch's own
-        // just-committed UPDATE from it.
-        let (h, a, _b) = rcsi_harness().await;
-        let reply = must_not_block(
-            &h,
-            a,
-            "UPDATE t SET v = 42 WHERE id = 1; \
-             DECLARE @x INT; SELECT @x = v FROM t WHERE id = 1; SELECT @x",
-        )
-        .await;
-        assert_eq!(values(&reply), vec![42]);
-    }
-
-    #[tokio::test]
-    async fn repeatable_read_still_blocks_under_rcsi() {
-        // RCSI changes READ COMMITTED only. A REPEATABLE READ reader keeps
-        // its Table S request and parks behind the writer.
-        let (h, a, b) = rcsi_harness().await;
-        h.handle
-            .run_batch(b, "SET TRANSACTION ISOLATION LEVEL REPEATABLE READ".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET v = 99 WHERE id = 1;".into())
-            .await
-            .unwrap();
-        let reader = {
-            let handle = h.handle.clone();
-            tokio::spawn(async move {
-                handle
-                    .run_batch(b, "SELECT v FROM t WHERE id = 1".into())
-                    .await
-            })
-        };
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = reader.await.unwrap().unwrap();
-        assert_eq!(
-            values(&reply),
-            vec![99],
-            "an RR reader sees only committed state, by waiting"
-        );
-    }
-
-    #[tokio::test]
-    async fn rcsi_heap_table_serves_snapshot_rows() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        for sql in [
-            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON",
-            "CREATE TABLE hp (v INT)",
-            "INSERT INTO hp VALUES (10), (20)",
-        ] {
-            h.handle.run_batch(a, sql.into()).await.unwrap();
-        }
-        h.handle
-            .run_batch(
-                a,
-                "BEGIN TRAN; UPDATE hp SET v = 99 WHERE v = 10; DELETE FROM hp WHERE v = 20;"
-                    .into(),
-            )
-            .await
-            .unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM hp").await;
-        let mut got = values(&reply);
-        got.sort();
-        assert_eq!(
-            got,
-            vec![10, 20],
-            "heap identities (RIDs) version like tree keys"
-        );
-        h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM hp").await;
-        let mut got = values(&reply);
-        got.sort();
-        assert_eq!(got, vec![10, 20]);
-    }
-
-    #[tokio::test]
-    async fn alter_database_rejects_wrong_names_and_open_transactions() {
-        let h = start(Duration::from_secs(30));
-        let s = h
-            .handle
-            .open_session("truthdb".into(), "sa".into(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let reply = h
-            .handle
-            .run_batch(
-                s,
-                "ALTER DATABASE nosuch SET READ_COMMITTED_SNAPSHOT ON".into(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), Some(911));
-        let reply = h
-            .handle
-            .run_batch(
-                s,
-                "BEGIN TRAN; ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT ON;".into(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), Some(226));
-        h.handle.run_batch(s, "ROLLBACK".into()).await.unwrap();
-        // The session's own database by name works, and the option sticks.
-        let reply = h
-            .handle
-            .run_batch(
-                s,
-                "ALTER DATABASE truthdb SET READ_COMMITTED_SNAPSHOT ON".into(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-    }
-
-    // The next six tests came out of the adversarial review: the first two
-    // pin its confirmed findings (both reproduced as dirty reads before the
-    // fixes), the rest are its passing probes kept as regressions.
-
-    #[tokio::test]
-    async fn serializable_exec_literal_select_must_not_dirty_read_under_rcsi() {
-        // A SERIALIZABLE session EXECs a literal SELECT while another
-        // session's transaction holds an uncommitted UPDATE. The reader must
-        // wait for the writer (SERIALIZABLE reads lock) and, after the
-        // writer's ROLLBACK, see the original value 10. A dirty read
-        // returns 99 — a value that never committed.
-        let (h, a, b) = rcsi_harness().await;
-        h.handle
-            .run_batch(b, "SET TRANSACTION ISOLATION LEVEL SERIALIZABLE".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET v = 99 WHERE id = 1;".into())
-            .await
-            .unwrap();
-        let reader = {
-            let handle = h.handle.clone();
-            tokio::spawn(async move {
-                handle
-                    .run_batch(
-                        b,
-                        "EXEC sp_executesql N'SELECT v FROM t WHERE id = 1'".into(),
-                    )
-                    .await
-            })
-        };
-        // Give the reader time to (incorrectly) run without blocking.
-        tokio::time::sleep(Duration::from_millis(300)).await;
-        h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-        let reply = reader.await.unwrap().unwrap();
-        assert_eq!(
-            values(&reply),
-            vec![10],
-            "a SERIALIZABLE reader must never see the uncommitted 99"
-        );
-    }
-
-    #[tokio::test]
-    async fn a_parked_reader_is_reanalyzed_when_rcsi_flips_off() {
-        // A reader analyzed while RCSI is ON (Database IS only, no Table S)
-        // parks behind a pending ALTER ... OFF; when granted, RCSI is OFF so
-        // it executes lock-based — with the stale (versioned) lock set. A
-        // concurrent uncommitted writer is then readable: dirty read at
-        // READ COMMITTED with RCSI off.
-        for round in 0..5 {
-            let (h, a, b) = rcsi_harness().await;
-            let c = h
-                .handle
-                .open_session(String::new(), String::new(), 0)
-                .await
-                .expect("open session")
-                .0;
-            // Writer A holds Database IX -> the ALTER (Database X) parks.
-            h.handle
-                .run_batch(a, "BEGIN TRAN; UPDATE t SET v = 55 WHERE id = 2;".into())
-                .await
-                .unwrap();
-            let admin = {
-                let handle = h.handle.clone();
-                let s = h
-                    .handle
-                    .open_session(String::new(), String::new(), 0)
-                    .await
-                    .expect("open session")
-                    .0;
-                tokio::spawn(async move {
-                    handle
-                        .run_batch(
-                            s,
-                            "ALTER DATABASE CURRENT SET READ_COMMITTED_SNAPSHOT OFF".into(),
-                        )
-                        .await
-                })
-            };
-            tokio::time::sleep(Duration::from_millis(100)).await;
-            // Writer C parks behind the ALTER (fairness on Database).
-            let writer = {
-                let handle = h.handle.clone();
-                tokio::spawn(async move {
-                    handle
-                        .run_batch(c, "BEGIN TRAN; UPDATE t SET v = 777 WHERE id = 1;".into())
-                        .await
-                })
-            };
-            tokio::time::sleep(Duration::from_millis(100)).await;
-            // Reader B analyzed NOW (RCSI still on -> no Table S), parks.
-            let reader = {
-                let handle = h.handle.clone();
-                tokio::spawn(async move {
-                    handle
-                        .run_batch(b, "SELECT v FROM t WHERE id = 1".into())
-                        .await
-                })
-            };
-            tokio::time::sleep(Duration::from_millis(100)).await;
-            // A commits: the ALTER runs and flips RCSI off. B's lock set was
-            // analyzed under RCSI-on (no Table S); the grant path must
-            // re-analyze it, so B now blocks behind C's open transaction
-            // instead of reading its uncommitted 777 lock-free.
-            h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-            // C's UPDATE batch completes (transaction stays open, holding
-            // its locks); B must still be waiting on them.
-            tokio::time::timeout(Duration::from_secs(5), writer)
-                .await
-                .expect("writer hung")
-                .unwrap()
-                .unwrap();
-            h.handle.run_batch(c, "ROLLBACK".into()).await.unwrap();
-            let reply = tokio::time::timeout(Duration::from_secs(5), reader)
-                .await
-                .expect("reader hung")
-                .unwrap()
-                .unwrap();
-            let seen = values(&reply);
-            assert_eq!(
-                seen,
-                vec![10],
-                "round {round}: the reader may only see committed state \
-                 (777 never committed — seeing it is the dirty read)"
-            );
-            admin.await.unwrap().unwrap();
-        }
-    }
-
-    #[tokio::test]
-    async fn rcsi_insert_after_delete_in_one_txn_serves_the_old_image() {
-        let (h, a, b) = rcsi_harness().await;
-        h.handle
-            .run_batch(
-                a,
-                "BEGIN TRAN; DELETE FROM t WHERE id = 1; INSERT INTO t VALUES (1, 111);".into(),
-            )
-            .await
-            .unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10], "pre-txn image, not the re-insert");
-        h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        h.handle
-            .run_batch(
-                a,
-                "BEGIN TRAN; DELETE FROM t WHERE id = 1; INSERT INTO t VALUES (1, 222);".into(),
-            )
-            .await
-            .unwrap();
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![222]);
-    }
-
-    #[tokio::test]
-    async fn rcsi_aggregates_and_subqueries_read_the_snapshot() {
-        let (h, a, b) = rcsi_harness().await;
-        h.handle
-            .run_batch(
-                a,
-                "BEGIN TRAN; DELETE FROM t WHERE id = 2; UPDATE t SET v = 99 WHERE id = 1;".into(),
-            )
-            .await
-            .unwrap();
-        // Aggregate over the collecting path (build_table_source).
-        let reply = must_not_block(&h, b, "SELECT COUNT(*) FROM t").await;
-        assert_eq!(
-            values(&reply),
-            vec![2],
-            "the uncommitted delete is invisible"
-        );
-        let reply = must_not_block(&h, b, "SELECT SUM(v) FROM t").await;
-        assert_eq!(values(&reply), vec![30], "10 + 20, not 99 + (deleted)");
-        // Uncorrelated subquery: reads t under the same statement snapshot.
-        let reply =
-            must_not_block(&h, b, "SELECT v FROM t WHERE id = (SELECT MAX(id) FROM t)").await;
-        assert_eq!(values(&reply), vec![20], "MAX(id) = 2 in the snapshot");
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT COUNT(*) FROM t").await;
-        assert_eq!(values(&reply), vec![1]);
-    }
-
-    #[tokio::test]
-    async fn rcsi_range_seek_does_not_double_count_a_moved_entry() {
-        let (h, a, b) = rcsi_harness().await;
-        let mut fill = String::from("INSERT INTO t VALUES ");
-        for i in 3..=25 {
-            fill.push_str(&format!("({}, {}),", i, i * 100));
-        }
-        fill.pop();
-        h.handle.run_batch(a, fill).await.unwrap();
-        h.handle
-            .run_batch(a, "CREATE INDEX ix_v ON t (v)".into())
-            .await
-            .unwrap();
-        // Uncommitted: moves id=1's entry from v=10 to v=15 — both inside
-        // the seek range below.
-        h.handle
-            .run_batch(a, "BEGIN TRAN; UPDATE t SET v = 15 WHERE id = 1;".into())
-            .await
-            .unwrap();
-        let reply = must_not_block(
-            &h,
-            b,
-            "SELECT id FROM t WHERE v >= 5 AND v <= 20 ORDER BY id",
-        )
-        .await;
-        assert_eq!(
-            values(&reply),
-            vec![1, 2],
-            "id=1 exactly once (image v=10), id=2 (v=20); no double count"
-        );
-        h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn rcsi_covering_seek_serves_a_deleted_row_image() {
-        let (h, a, b) = rcsi_harness().await;
-        let mut fill = String::from("INSERT INTO t VALUES ");
-        for i in 3..=25 {
-            fill.push_str(&format!("({}, {}),", i, i * 100));
-        }
-        fill.pop();
-        h.handle.run_batch(a, fill).await.unwrap();
-        h.handle
-            .run_batch(a, "CREATE INDEX ix_vc ON t (v) INCLUDE (v, id)".into())
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "BEGIN TRAN; DELETE FROM t WHERE id = 1;".into())
-            .await
-            .unwrap();
-        // The deleted row's index entry is gone; the covering seek must
-        // append its image from the version store.
-        let reply = must_not_block(&h, b, "SELECT id FROM t WHERE v = 10").await;
-        assert_eq!(values(&reply), vec![1], "the deleted row's image is served");
-        h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT id FROM t WHERE v = 10").await;
-        assert_eq!(values(&reply), Vec::<i64>::new());
-    }
-
-    // ---- SNAPSHOT isolation (Stage 13) -----------------------------------
-
-    /// ALLOW_SNAPSHOT_ISOLATION on (RCSI deliberately off — SNAPSHOT must
-    /// stand alone), `t (id PK, v)` = (1,10), (2,20); session B is SNAPSHOT.
-    async fn si_harness() -> (Harness, SessionId, SessionId) {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        for sql in [
-            "ALTER DATABASE CURRENT SET ALLOW_SNAPSHOT_ISOLATION ON",
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            "INSERT INTO t VALUES (1, 10), (2, 20)",
-        ] {
-            let reply = h.handle.run_batch(a, sql.into()).await.unwrap();
-            assert_eq!(
-                error_number(&reply),
-                None,
-                "{sql}: {:?}",
-                reply.outcome.error
-            );
-        }
-        h.handle
-            .run_batch(b, "SET TRANSACTION ISOLATION LEVEL SNAPSHOT".into())
-            .await
-            .unwrap();
-        (h, a, b)
-    }
-
-    #[tokio::test]
-    async fn snapshot_reads_are_repeatable_and_block_nobody() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(
-            values(&reply),
-            vec![10],
-            "first access establishes the view"
-        );
-        // A writer neither blocks the snapshot reader nor is blocked by it.
-        let reply = must_not_block(&h, a, "UPDATE t SET v = 99 WHERE id = 1").await;
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(
-            values(&reply),
-            vec![10],
-            "the committed 99 postdates the transaction's snapshot — repeatable read"
-        );
-        h.handle.run_batch(b, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(
-            values(&reply),
-            vec![99],
-            "a new transaction sees the new state"
-        );
-    }
-
-    #[tokio::test]
-    async fn snapshot_autocommit_statements_take_fresh_snapshots() {
-        let (h, a, b) = si_harness().await;
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        h.handle
-            .run_batch(a, "UPDATE t SET v = 99 WHERE id = 1".into())
-            .await
-            .unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(
-            values(&reply),
-            vec![99],
-            "outside a transaction each statement is its own snapshot"
-        );
-    }
-
-    #[tokio::test]
-    async fn snapshot_update_conflict_is_3960_and_rolls_the_transaction_back() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        // A commits a change B's snapshot cannot see...
-        h.handle
-            .run_batch(a, "UPDATE t SET v = 99 WHERE id = 1".into())
-            .await
-            .unwrap();
-        // ...so B writing the same row is the classic update conflict.
-        let reply = h
-            .handle
-            .run_batch(b, "UPDATE t SET v = 100 WHERE id = 1".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3960),
-            "{:?}",
-            reply.outcome.error
-        );
-        // SQL Server ABORTS the transaction (not merely dooms it): a COMMIT
-        // now has nothing to commit.
-        let reply = h.handle.run_batch(b, "COMMIT".into()).await.unwrap();
-        assert_eq!(error_number(&reply), Some(3902), "the transaction is gone");
-        // The conflicting write never landed; A's did.
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![99]);
-    }
-
-    #[tokio::test]
-    async fn snapshot_delete_of_a_row_deleted_since_the_snapshot_is_3960() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 2").await;
-        assert_eq!(values(&reply), vec![20]);
-        h.handle
-            .run_batch(a, "DELETE FROM t WHERE id = 2".into())
-            .await
-            .unwrap();
-        // The row is physically gone, but B's snapshot still sees it — and
-        // targeting it must conflict, not silently affect zero rows.
-        let reply = h
-            .handle
-            .run_batch(b, "DELETE FROM t WHERE id = 2".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3960),
-            "{:?}",
-            reply.outcome.error
-        );
-    }
-
-    #[tokio::test]
-    async fn snapshot_dml_targets_snapshot_rows_not_current_ones() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        // A moves the row INTO the predicate's range after B's snapshot.
-        h.handle
-            .run_batch(a, "UPDATE t SET v = 77 WHERE id = 1".into())
-            .await
-            .unwrap();
-        // B's WHERE v = 77 matches the CURRENT row but not the snapshot's
-        // version: nothing is targeted, no conflict — current-based targeting
-        // would wrongly update it.
-        let reply = h
-            .handle
-            .run_batch(b, "UPDATE t SET v = 1000 WHERE v = 77".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        match reply.outcome.results.as_slice() {
-            [StatementResult::RowsAffected(n)] => {
-                assert_eq!(*n, 0, "the snapshot's rows decide targeting")
-            }
-            other => panic!("expected RowsAffected, got {other:?}"),
-        }
-        h.handle.run_batch(b, "ROLLBACK".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![77], "A's committed value is untouched");
-    }
-
-    #[tokio::test]
-    async fn snapshot_without_allow_option_is_3952_at_access() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(
-                a,
-                "CREATE TABLE t (id INT NOT NULL PRIMARY KEY); INSERT INTO t VALUES (1);".into(),
-            )
-            .await
-            .unwrap();
-        // The SET itself succeeds (SQL Server defers the check to access).
-        let reply = h
-            .handle
-            .run_batch(a, "SET TRANSACTION ISOLATION LEVEL SNAPSHOT".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        let reply = h
-            .handle
-            .run_batch(a, "SELECT id FROM t".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3952),
-            "{:?}",
-            reply.outcome.error
-        );
-        // Inside a transaction the failure dooms it: COMMIT is refused until
-        // the rollback.
-        h.handle.run_batch(a, "BEGIN TRAN".into()).await.unwrap();
-        let reply = h
-            .handle
-            .run_batch(a, "SELECT id FROM t".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), Some(3952));
-        let reply = h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3930),
-            "doomed: only ROLLBACK is allowed"
-        );
-        let reply = h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-    }
-
-    #[tokio::test]
-    async fn snapshot_schema_change_since_the_snapshot_is_3961() {
-        let (h, a, b) = si_harness().await;
-        h.handle
-            .run_batch(a, "CREATE TABLE other (id INT NOT NULL PRIMARY KEY)".into())
-            .await
-            .unwrap();
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        // The ALTER re-encodes every row of t; B's images predate it.
-        h.handle
-            .run_batch(a, "ALTER TABLE t ADD extra INT DEFAULT 5".into())
-            .await
-            .unwrap();
-        let reply = h
-            .handle
-            .run_batch(b, "SELECT v FROM t WHERE id = 1".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3961),
-            "{:?}",
-            reply.outcome.error
-        );
-        // 3961 fails the statement, not the transaction: other tables still
-        // read, and the transaction can end normally.
-        let reply = must_not_block(&h, b, "SELECT id FROM other").await;
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        let reply = h.handle.run_batch(b, "COMMIT".into()).await.unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-    }
-
-    #[tokio::test]
-    async fn snapshot_inserts_do_not_conflict_and_stay_isolated() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT id FROM t").await;
-        let mut got = values(&reply);
-        got.sort();
-        assert_eq!(got, vec![1, 2]);
-        h.handle
-            .run_batch(a, "INSERT INTO t VALUES (3, 30)".into())
-            .await
-            .unwrap();
-        // B's own insert succeeds (insert-insert is a key collision question,
-        // not a version conflict) and B still does not see A's row.
-        let reply = must_not_block(&h, b, "INSERT INTO t VALUES (4, 40)").await;
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        let reply = must_not_block(&h, b, "SELECT id FROM t").await;
-        let mut got = values(&reply);
-        got.sort();
-        assert_eq!(
-            got,
-            vec![1, 2, 4],
-            "own insert visible, A's post-snapshot one not"
-        );
-        h.handle.run_batch(b, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT id FROM t").await;
-        let mut got = values(&reply);
-        got.sort();
-        assert_eq!(got, vec![1, 2, 3, 4]);
-    }
-
-    // ---- SI review PoC tests ---------------------------------------------
-
-    /// Attack A: a row re-keyed by a snapshot-invisible writer. Targeting the
-    /// snapshot row by its OLD key must be 3960 (its current state was
-    /// produced by an invisible writer); targeting the NEW key must affect
-    /// nothing (the snapshot has no such row) and must not touch the current
-    /// row.
-    #[tokio::test]
-    async fn snapshot_rekeyed_row_conflicts_on_old_key_and_misses_on_new_key() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 2").await;
-        assert_eq!(values(&reply), vec![20]);
-        // A re-keys row 2 -> 5 after B's snapshot.
-        let reply = h
-            .handle
-            .run_batch(a, "UPDATE t SET id = 5 WHERE id = 2".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        // Targeting by the NEW key: the snapshot has no row with id = 5.
-        let reply = h
-            .handle
-            .run_batch(b, "DELETE FROM t WHERE id = 5".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        match reply.outcome.results.as_slice() {
-            [StatementResult::RowsAffected(n)] => assert_eq!(*n, 0, "snapshot has no id = 5"),
-            other => panic!("expected RowsAffected, got {other:?}"),
-        }
-        // Targeting by the OLD key: the snapshot row exists but its current
-        // state was produced by an invisible writer - update conflict.
-        let reply = h
-            .handle
-            .run_batch(b, "UPDATE t SET v = 200 WHERE id = 2".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3960),
-            "{:?}",
-            reply.outcome.error
-        );
-        // The 3960 rolled B back; the current row was never touched by the
-        // no-op DELETE or the conflicting UPDATE.
-        let reply = must_not_block(&h, a, "SELECT v FROM t WHERE id = 5").await;
-        assert_eq!(values(&reply), vec![20], "A's re-keyed row survived");
-        let reply = must_not_block(&h, a, "SELECT COUNT(*) FROM t").await;
-        assert_eq!(values(&reply), vec![2]);
-    }
-
-    /// Attack E: an index created AFTER the snapshot only contains
-    /// post-snapshot states. A seek through it must still produce exactly the
-    /// snapshot's rows: old values found via chain images, new values
-    /// filtered out.
-    #[tokio::test]
-    async fn snapshot_reads_merge_through_an_index_created_after_the_snapshot() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        // Post-snapshot: an index on v, then an update that moves row 1's
-        // entry from 10 to 99 (the new index only ever contains 99).
-        let reply = h
-            .handle
-            .run_batch(a, "CREATE INDEX ix_v ON t (v)".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        let reply = h
-            .handle
-            .run_batch(a, "UPDATE t SET v = 99 WHERE id = 1".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        // Seek for the snapshot's value: nothing physical in range, so the
-        // row must come from its chain image.
-        let reply = must_not_block(&h, b, "SELECT id FROM t WHERE v = 10").await;
-        assert_eq!(
-            values(&reply),
-            vec![1],
-            "the snapshot's v = 10 row must be found through the new index"
-        );
-        // Seek for the current value: the physical entry resolves to an
-        // image with v = 10, which the predicate must filter out.
-        let reply = must_not_block(&h, b, "SELECT id FROM t WHERE v = 99").await;
-        assert_eq!(
-            values(&reply),
-            Vec::<i64>::new(),
-            "the current v = 99 row postdates the snapshot"
-        );
-        h.handle.run_batch(b, "COMMIT".into()).await.unwrap();
-    }
-
-    /// Attack C: EXEC of a literal shares the transaction's snapshot - the
-    /// inner statement both CAPTURES it (first access inside an EXEC) and
-    /// REUSES it, and 3952 fires at the inner statement's access when the
-    /// option is off.
-    #[tokio::test]
-    async fn exec_literal_shares_the_transaction_snapshot() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        // First data access happens INSIDE the EXEC: it must establish the
-        // transaction's snapshot.
-        let reply =
-            must_not_block(&h, b, "EXEC sp_executesql N'SELECT v FROM t WHERE id = 1'").await;
-        assert_eq!(values(&reply), vec![10]);
-        h.handle
-            .run_batch(a, "UPDATE t SET v = 99 WHERE id = 1".into())
-            .await
-            .unwrap();
-        // Both a plain statement and another EXEC read the SAME view.
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(
-            values(&reply),
-            vec![10],
-            "captured inside EXEC, reused outside"
-        );
-        let reply =
-            must_not_block(&h, b, "EXEC sp_executesql N'SELECT v FROM t WHERE id = 1'").await;
-        assert_eq!(values(&reply), vec![10], "reused inside EXEC");
-        h.handle.run_batch(b, "COMMIT".into()).await.unwrap();
-    }
-
-    /// Attack C/D: without ALLOW_SNAPSHOT_ISOLATION, the 3952 fires at the
-    /// INNER statement of an EXEC too, and dooms the open transaction.
-    #[tokio::test]
-    async fn exec_literal_under_snapshot_without_allow_is_3952() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(
-                a,
-                "CREATE TABLE t (id INT NOT NULL PRIMARY KEY); INSERT INTO t VALUES (1);".into(),
-            )
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "SET TRANSACTION ISOLATION LEVEL SNAPSHOT".into())
-            .await
-            .unwrap();
-        h.handle.run_batch(a, "BEGIN TRAN".into()).await.unwrap();
-        let reply = h
-            .handle
-            .run_batch(a, "EXEC sp_executesql N'SELECT id FROM t'".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3952),
-            "{:?}",
-            reply.outcome.error
-        );
-        let reply = h.handle.run_batch(a, "COMMIT".into()).await.unwrap();
-        assert_eq!(error_number(&reply), Some(3930), "doomed");
-        let reply = h.handle.run_batch(a, "ROLLBACK".into()).await.unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        // The session is usable again (autocommit statements still 3952
-        // until the level changes, but the SET path works).
-        h.handle
-            .run_batch(a, "SET TRANSACTION ISOLATION LEVEL READ COMMITTED".into())
-            .await
-            .unwrap();
-        let reply = must_not_block(&h, a, "SELECT id FROM t").await;
-        assert_eq!(values(&reply), vec![1]);
-    }
-
-    /// Attack B / leak audit: nested BEGIN and a savepoint rollback keep the
-    /// transaction - and must keep its snapshot.
-    #[tokio::test]
-    async fn nested_begin_and_savepoint_rollback_keep_the_snapshot() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        h.handle
-            .run_batch(a, "UPDATE t SET v = 99 WHERE id = 1".into())
-            .await
-            .unwrap();
-        // Nested BEGIN + inner COMMIT: trancount 2 -> 1, transaction alive.
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        h.handle.run_batch(b, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(
-            values(&reply),
-            vec![10],
-            "inner COMMIT must not drop the view"
-        );
-        // Savepoint rollback: own write undone, transaction and view alive.
-        h.handle.run_batch(b, "SAVE TRAN sp1".into()).await.unwrap();
-        let reply = h
-            .handle
-            .run_batch(b, "UPDATE t SET v = 500 WHERE id = 2".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        h.handle
-            .run_batch(b, "ROLLBACK TRAN sp1".into())
-            .await
-            .unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 2").await;
-        assert_eq!(
-            values(&reply),
-            vec![20],
-            "own write rolled back to the savepoint"
-        );
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(
-            values(&reply),
-            vec![10],
-            "the view survives the savepoint rollback"
-        );
-        h.handle.run_batch(b, "COMMIT".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![99]);
-    }
-
-    /// Attack E (option toggle): ALTER DATABASE SET ALLOW_SNAPSHOT_ISOLATION
-    /// OFF while a SNAPSHOT transaction holds a registered snapshot between
-    /// batches (it holds no locks then, so Database X does not exclude it).
-    /// Expected correct behavior: the toggle succeeds (or waits), the
-    /// snapshot transaction gets 3952 at its next access, and the session
-    /// recovers. Today the version store's reset debug_asserts that no
-    /// snapshot is live - which is false here.
-    #[tokio::test]
-    async fn alter_options_refuse_while_a_snapshot_transaction_lives() {
-        // A SNAPSHOT transaction idle between batches holds no locks, so the
-        // ALTER's Database X grants — but flipping the options under its
-        // registered snapshot would reset the store its reads depend on
-        // (the review's HIGH: a debug panic, silent wrong data in release).
-        // The ALTER refuses with 5061 instead; SQL Server would wait.
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10], "snapshot registered");
-        let reply = tokio::time::timeout(
-            Duration::from_secs(5),
-            h.handle.run_batch(
-                a,
-                "ALTER DATABASE CURRENT SET ALLOW_SNAPSHOT_ISOLATION OFF".into(),
-            ),
-        )
-        .await
-        .expect("the ALTER must not hang")
-        .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(5061),
-            "refused while the snapshot lives: {:?}",
-            reply.outcome.error
-        );
-        // The transaction is untouched — still repeatable.
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        h.handle.run_batch(b, "COMMIT".into()).await.unwrap();
-        // Snapshot released: the retry succeeds, and SNAPSHOT access is gone.
-        let reply = h
-            .handle
-            .run_batch(
-                a,
-                "ALTER DATABASE CURRENT SET ALLOW_SNAPSHOT_ISOLATION OFF".into(),
-            )
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        let reply = h
-            .handle
-            .run_batch(b, "SELECT v FROM t WHERE id = 1".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), Some(3952));
-    }
-
-    /// A statement-scoped (autocommit) 3952 does not wedge the session.
-    #[tokio::test]
-    async fn autocommit_3952_leaves_the_session_usable() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        h.handle
-            .run_batch(
-                a,
-                "CREATE TABLE t (id INT NOT NULL PRIMARY KEY); INSERT INTO t VALUES (1);".into(),
-            )
-            .await
-            .unwrap();
-        h.handle
-            .run_batch(a, "SET TRANSACTION ISOLATION LEVEL SNAPSHOT".into())
-            .await
-            .unwrap();
-        let reply = h
-            .handle
-            .run_batch(a, "SELECT id FROM t".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), Some(3952));
-        // A table-free SELECT is not data access: no 3952, and it does not
-        // establish a snapshot (SQL Server defers both to the first read of
-        // an actual object) — the review's finding 3.
-        let reply = h.handle.run_batch(a, "SELECT 1".into()).await.unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        h.handle
-            .run_batch(a, "SET TRANSACTION ISOLATION LEVEL READ COMMITTED".into())
-            .await
-            .unwrap();
-        let reply = must_not_block(&h, a, "SELECT id FROM t").await;
-        assert_eq!(values(&reply), vec![1]);
-    }
-
-    /// 3961 completeness: DROP TABLE + CREATE TABLE of the same name between
-    /// a SNAPSHOT transaction's batches. SQL Server raises 3961 (metadata is
-    /// not versioned); at minimum the reader must not silently see the NEW
-    /// table's contents as if they were its snapshot.
-    #[tokio::test]
-    async fn snapshot_drop_and_recreate_between_batches_is_3961() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        for sql in [
-            "DROP TABLE t",
-            "CREATE TABLE t (id INT NOT NULL PRIMARY KEY, v INT)",
-            "INSERT INTO t VALUES (7, 70)",
-        ] {
-            let reply = h.handle.run_batch(a, sql.into()).await.unwrap();
-            assert_eq!(
-                error_number(&reply),
-                None,
-                "{sql}: {:?}",
-                reply.outcome.error
-            );
-        }
-        let reply = h
-            .handle
-            .run_batch(b, "SELECT v FROM t".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3961),
-            "SQL Server: DDL since the snapshot is 3961, got {:?} / rows {:?}",
-            reply.outcome.error,
-            reply.outcome.results
-        );
-        h.handle.run_batch(b, "ROLLBACK".into()).await.unwrap();
-    }
-
-    /// The heap arm of the SI DML target scan: conflicts through synthesized
-    /// RID locators (no primary key anywhere).
-    #[tokio::test]
-    async fn snapshot_heap_dml_conflicts_via_rid_identities() {
-        let h = start(Duration::from_secs(30));
-        let a = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        let b = h
-            .handle
-            .open_session(String::new(), String::new(), 0)
-            .await
-            .expect("open session")
-            .0;
-        for sql in [
-            "ALTER DATABASE CURRENT SET ALLOW_SNAPSHOT_ISOLATION ON",
-            "CREATE TABLE hp (v INT)",
-            "INSERT INTO hp VALUES (10), (20)",
-        ] {
-            let reply = h.handle.run_batch(a, sql.into()).await.unwrap();
-            assert_eq!(
-                error_number(&reply),
-                None,
-                "{sql}: {:?}",
-                reply.outcome.error
-            );
-        }
-        h.handle
-            .run_batch(b, "SET TRANSACTION ISOLATION LEVEL SNAPSHOT".into())
-            .await
-            .unwrap();
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM hp WHERE v = 10").await;
-        assert_eq!(values(&reply), vec![10]);
-        // A deletes the row B's snapshot still sees.
-        let reply = h
-            .handle
-            .run_batch(a, "DELETE FROM hp WHERE v = 10".into())
-            .await
-            .unwrap();
-        assert_eq!(error_number(&reply), None, "{:?}", reply.outcome.error);
-        // Targeting it via its synthesized RID locator is the conflict.
-        let reply = h
-            .handle
-            .run_batch(b, "DELETE FROM hp WHERE v = 10".into())
-            .await
-            .unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3960),
-            "{:?}",
-            reply.outcome.error
-        );
-        // B was rolled back; the surviving row is intact.
-        let reply = must_not_block(&h, b, "SELECT v FROM hp").await;
-        assert_eq!(values(&reply), vec![20]);
-    }
-
-    /// Full-table DML under SI: DELETE with no WHERE keeps every snapshot
-    /// row; if ANY of them was changed since the snapshot, 3960.
-    #[tokio::test]
-    async fn snapshot_full_table_delete_conflicts_when_any_row_changed() {
-        let (h, a, b) = si_harness().await;
-        h.handle.run_batch(b, "BEGIN TRAN".into()).await.unwrap();
-        let reply = must_not_block(&h, b, "SELECT v FROM t WHERE id = 1").await;
-        assert_eq!(values(&reply), vec![10]);
-        h.handle
-            .run_batch(a, "UPDATE t SET v = 99 WHERE id = 2".into())
-            .await
-            .unwrap();
-        let reply = h.handle.run_batch(b, "DELETE FROM t".into()).await.unwrap();
-        assert_eq!(
-            error_number(&reply),
-            Some(3960),
-            "{:?}",
-            reply.outcome.error
-        );
-    }
-}
+mod tests;
